% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-basic.R
\name{abs_list}
\alias{abs_list}
\alias{abs_list,Cycif-method}
\alias{abs_list,CycifStack-method}
\title{Get the list of antibodies used in a CycifStack object}
\usage{
abs_list(x)

\S4method{abs_list}{Cycif}(x)

\S4method{abs_list}{CycifStack}(x)
}
\arguments{
\item{x}{A CycifStack object.}
}
\value{
A data frame containing the absolute list of markers with columns "ab" and "cycle."
}
\description{
The \code{abs_list} function returns the absolute list of markers in a CycifStack object.
}
\seealso{
\code{\link{nSamples}}: Get the number of samples in a CycifStack object.
\code{\link{maxCycles}}: Get the maximum number of cycles in a CycifStack object.
\code{\link{within_rois}}: Check if samples in a CycifStack are within ROIs.
\code{\link{length}}: Equivalent to \code{nSamples} for CycifStack objects.
}
