% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-cell-neighborhood.R
\name{cell_neighborhood}
\alias{cell_neighborhood}
\alias{cell_neighborhood,Cycif-method}
\alias{cell_neighborhood,CycifStack-method}
\alias{cell_neighborhood<-}
\alias{cell_neighborhood<-,Cycif-method}
\alias{cell_neighborhood<-,CycifStack-method}
\title{Get/set the CellNeighborhood object on a Cycif or CycifStack}
\usage{
cell_neighborhood(x, ...)

\S4method{cell_neighborhood}{Cycif}(x)

\S4method{cell_neighborhood}{CycifStack}(x)

cell_neighborhood(x) <- value

\S4method{cell_neighborhood}{Cycif}(x) <- value

\S4method{cell_neighborhood}{CycifStack}(x) <- value
}
\arguments{
\item{x}{A Cycif or CycifStack object.}

\item{...}{Additional arguments (unused).}

\item{value}{A CellNeighborhood object (for the setter).}
}
\description{
Get/set the CellNeighborhood object on a Cycif or CycifStack
}
