% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-cell-neighborhood.R
\name{neighborCounts}
\alias{neighborCounts}
\alias{neighborCounts,ANY-method}
\alias{neighborFreq}
\alias{neighborFreq,ANY-method}
\alias{neighborDensity}
\alias{neighborDensity,ANY-method}
\alias{meanExprPerNbhd}
\alias{meanExprPerNbhd,ANY-method}
\alias{meanExprPerCtNbhd}
\alias{meanExprPerCtNbhd,ANY-method}
\alias{distToTumorBorder}
\alias{distToTumorBorder,ANY-method}
\alias{neighborGraph}
\alias{neighborGraph,ANY-method}
\title{Convenience accessors for CellNeighborhood fields on a Cycif or CycifStack}
\usage{
neighborCounts(x)

\S4method{neighborCounts}{ANY}(x)

neighborFreq(x)

\S4method{neighborFreq}{ANY}(x)

neighborDensity(x)

\S4method{neighborDensity}{ANY}(x)

meanExprPerNbhd(x)

\S4method{meanExprPerNbhd}{ANY}(x)

meanExprPerCtNbhd(x)

\S4method{meanExprPerCtNbhd}{ANY}(x)

distToTumorBorder(x)

\S4method{distToTumorBorder}{ANY}(x)

neighborGraph(x)

\S4method{neighborGraph}{ANY}(x)
}
\arguments{
\item{x}{A Cycif or CycifStack object with a CellNeighborhood already computed.}
}
\description{
Each of these is a thin wrapper around \code{cell_neighborhood(x)@<slot>}.
All are sized \code{sum(is_used(cell_neighborhood(x)))} rows, in the same
order -- see \code{\linkS4class{CellNeighborhood}}.
}
