% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype-accessors.R
\name{cell_types}
\alias{cell_types}
\alias{cell_types,CellTypes-method}
\alias{cell_types,Cycif-method}
\alias{cell_types,CycifStack-method}
\title{Get cell type information from a CellTypes, Cycif, or CycifStack object.}
\usage{
cell_types(x, ...)

\S4method{cell_types}{CellTypes}(x, strict = FALSE)

\S4method{cell_types}{Cycif}(x, ct_name = "default", strict = FALSE)

\S4method{cell_types}{CycifStack}(x, ct_name = "default", strict = FALSE)
}
\arguments{
\item{x}{A CellTypes, Cycif, or CycifStack object.}

\item{...}{Additional arguments (unused).}

\item{strict}{Logical, whether to include only strict cell type assignments.
Strict assignment returns NA when more than one cell type can be assigned to a cell, e.g., CD8T cell and Tumor cell  (default is FALSE).}

\item{ct_name}{Name of the cell type calling method (default is "default").}
}
\value{
A data frame with cell type information, including sample names and cell types.
}
\description{
This function retrieves cell type information from a CellTypes, Cycif, or CycifStack object.
You can specify the cell type calling method to use (if applicable) and whether
to include only strict cell type assignments.
}
\seealso{
\code{\link{Cycif}}, \code{\link{CycifStack}}, \code{\linkS4class{CellTypes}}
}
