% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{CellTypeCalling}
\alias{CellTypeCalling}
\title{Cell type calling function}
\usage{
CellTypeCalling(cy,p_thres=0.5,ctype.full=TRUE)
}
\arguments{
\item{cy}{A cycif obj.}

\item{p_thres}{A numerical between 0 and 1. A fixed probability that corresponds to the threshold intensity.}

\item{ctype.full}{logical. if cell types with fuller definition (e.g., stratification with PD-1/PD-L1 status) should be used.}
}
\description{
Cell type calling function
}
