% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellType-class.R
\docType{class}
\name{CellTypeCycif-class}
\alias{CellTypeCycif-class}
\title{A class that defines cell types in Cycif class, extending CellTypeDef class}
\description{
A class that defines cell types in Cycif class, extending CellTypeDef class
}
