% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypeCycifStack-utils.R
\name{nCycles,CellTypeCycifStack-method}
\alias{nCycles,CellTypeCycifStack-method}
\alias{uniq_abs<-}
\title{Accessing slots of a CellTypeCycifStack object}
\usage{
\S4method{nCycles}{CellTypeCycifStack}(x)

uniq_abs(x, ...) <- value
}
\arguments{
\item{x}{A CellTypeCycifStack object}

\item{value}{A character vector containing antibodies as in column names of
the protein expression matrix.}
}
\description{
Accessing slots of a CellTypeCycifStack object
}
