% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellType-class.R
\docType{class}
\name{CellTypeDef-class}
\alias{CellTypeDef-class}
\title{A class that defines cell types and antibodies in a CyCIF project}
\description{
A class that defines cell types and antibodies in a CyCIF project
}
