% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypeDef.R
\name{CellTypeDef}
\alias{CellTypeDef}
\alias{show,CellTypeDef-method}
\title{Instanciate and show a CellTypeDef object}
\usage{
CellTypeDef(filename)

\S4method{show}{CellTypeDef}(object)
}
\arguments{
\item{filename}{Excel sheet that contains two spreadsheets, named 'cell lineage' and 'cell state', respectively.}

\item{object}{A CellTypeDef object}
}
\description{
Instanciate and show a CellTypeDef object
}
