% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{CellTypeDefault-class}
\alias{CellTypeDefault-class}
\title{A class that defines cell types and antibodies in a CyCIF project (defunct)}
\description{
A class that defines cell types and antibodies in a CyCIF project (defunct)
}
