% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellType-utils.R
\name{cellTypes}
\alias{cellTypes}
\alias{markers}
\alias{cell_lineages}
\alias{cell_lineage_df}
\alias{cell_state_df}
\title{Accessing slots in CellType* class}
\usage{
cellTypes(x)

markers(x)

cell_lineages(x)

cell_lineage_df(x)

cell_state_df(x)
}
\arguments{
\item{x}{A CellTypeDefault object}
}
\description{
Accessing slots in CellType* class
}
