% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{CellTypes-class}
\alias{CellTypes-class}
\title{A class that defines cell types in Cycif class, extending CellTypeDefault class}
\description{
A class that defines cell types in Cycif class, extending CellTypeDefault class
}
