% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters-core.R
\name{check_abs}
\alias{check_abs}
\alias{check_abs,CycifStack-method}
\alias{set_abs<-}
\alias{set_abs<-,Cycif-method}
\alias{set_abs<-,CycifStack-method}
\title{Check and set antibodies for a Cycif or CycifStack object}
\usage{
check_abs(x)

\S4method{check_abs}{CycifStack}(x)

set_abs(x, ...) <- value

\S4method{set_abs}{Cycif}(x) <- value

\S4method{set_abs}{CycifStack}(x) <- value
}
\arguments{
\item{x}{A CycifStack object.}

\item{...}{Additional arguments (unused; for \verb{set_abs<-}).}

\item{value}{A character vector containing the new set of antibodies (for \verb{set_abs<-}).}
}
\value{
For \code{check_abs}, a warning message is displayed if the samples do not have the same
number of antibodies or if the antibody names are not identical.

For \verb{set_abs<-}, a modified CycifStack object with the specified antibodies is returned.
}
\description{
The \code{check_abs} function checks whether all samples in a CycifStack object have the same number
of antibodies and whether they have identical antibody names. If not, it provides
suggestions on how to address the issue.

The \code{set_abs} function allows you to set the
antibodies for a CycifStack object to a specified set of antibodies.
}
