% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis.R
\name{computeRCN}
\alias{computeRCN}
\alias{computeRCN,Cycif-method}
\alias{computeRCN,CycifStack-method}
\title{Compute Recurrent Cell Neighbors (RCN) for Cycif or CycifStack Objects}
\usage{
computeRCN(x, ...)

\S4method{computeRCN}{Cycif}(
  x,
  r,
  unit = c("pixel", "um"),
  focused.cts,
  cts.in.rcn,
  n.sampling,
  seed = 123
)

\S4method{computeRCN}{CycifStack}(
  x,
  r,
  unit = c("pixel", "um"),
  focused.cts,
  cts.in.rcn,
  n.sampling,
  seed = 123
)
}
\arguments{
\item{x}{A Cycif or CycifStack object.}

\item{r}{The radius within which neighboring cells are considered (in 'unit').}

\item{unit}{The unit of measurement for the radius ('pixel' or 'um'). If 'um' is specified,
the radius 'r' will be converted to pixels based on the assumed resolution (0.65 um per pixel).}

\item{focused.cts}{A character vector specifying the cell types of interest for RCN analysis.}

\item{cts.in.rcn}{A character vector specifying the cell types to consider when computing RCN values.}

\item{n.sampling}{The number of cells to randomly sample for RCN analysis.}

\item{seed}{The random seed for reproducibility.}
}
\value{
A list containing the following components:
\itemize{
\item 'within.rois': A logical vector indicating whether each cell is within a region of interest (ROI).
\item 'focused.cts': A character vector specifying the cell types of interest for RCN analysis.
\item 'cts.in.rcn': A character vector specifying the cell types considered when computing RCN values.
\item 'n.cells.selected': The number of cells selected for RCN analysis.
\item 'frnn': A list with recurrent Neighborhood information, including 'dist', 'id', 'eps', and 'sort'.
\item 'exp': A data frame containing expression data for selected cells.
\item 'is.selected': A logical vector indicating whether each cell is selected for RCN analysis.
\item 'rcn.count': A data frame containing the counts of neighboring cell types.
\item 'rcn.freq': A data frame containing the relative frequencies of neighboring cell types.
}
}
\description{
This function computes the Recurrent Cell Neighbors (RCN) for Cycif or CycifStack objects.
RCN measures the relative frequencies of neighboring cell types around each cell within specified
radius 'r'. The RCN analysis can be performed on a single Cycif object or across a CycifStack object.
}
\seealso{
\code{\link{cyApply}}, \code{\link{dbscan::frNN}}
}
