% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{createRSC}
\alias{createRSC}
\title{Create RowSideColors for Heatmap Visualization (TALAVE)}
\usage{
createRSC(cs, with.clusters = T, ld_name, order)
}
\arguments{
\item{cs}{A \code{CycifStack} object or other data structure containing sample annotations used for color assignment.}

\item{with.clusters}{Logical value indicating whether to include clustering information in the RowSideColors (default is TRUE). If TRUE, the function expects a \code{ld_name} to specify the clustering to be used.}

\item{ld_name}{A character string specifying the name of the UMAP or clustering to be used when \code{with.clusters} is TRUE. Required if \code{with.clusters} is TRUE.}

\item{order}{A character string specifying the desired order of the RowSideColors. The order can be a combination of the following letters: "B" for BOR, "c" for clusters, "T" for TimePoint, "g" for gBRCA.status, and "P" for Patient.ID. This parameter allows you to customize the order of the RowSideColors according to your preference (default is NULL).}
}
\value{
A list containing the following elements:
\itemize{
\item \code{pd}: A data frame with row names and the specified annotations.
\item \code{rsc}: A matrix representing the RowSideColors for the heatmap.
\item \code{row.order}: The order of rows in the RSC matrix if custom ordering is applied (only returned if the \code{order} parameter is provided).
}
}
\description{
This function generates RowSideColors (RSC) for use in heatmap visualization, typically with the \code{heatmap3} or \code{h3} functions. RSC assigns colors to rows (e.g., samples or features) based on sample annotations, such as Patient ID, gBRCA mutation status, time points, BOR (type of breast cancer), and clusters (for dimension reduction analysis).
}
\details{
The \code{createRSC} function generates RowSideColors for heatmap visualization based on sample annotations. The function assigns colors to rows (samples or features) in the heatmap, making it easier to identify patterns and relationships between samples.

The available annotations for color assignment are:
\itemize{
\item Patient ID
\item gBRCA mutation status (MUT or WT)
\item Time point
\item BOR (type of breast cancer)
\item Clusters (dimension reduction clusters, requires specifying \code{ld_name} when \code{with.clusters} is TRUE)
}

You can customize the order of the RowSideColors using the \code{order} parameter, which allows you to specify the order of annotations to be displayed in the heatmap from left to right. For example, setting \code{order = "BPcgT"} will order the annotations as BOR, Patient ID, clusters, gBRCA status, and TimePoint.

This function is especially useful for visualizing multi-dimensional data, such as cytometry data, in a heatmap, allowing you to explore relationships between samples and their associated annotations.
}
