% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype-accessors.R
\name{ct_names}
\alias{ct_names}
\alias{ct_names,Cycif-method}
\alias{ct_names,CycifStack-method}
\title{Get the names of cell types in a Cycif or CycifStack object}
\usage{
ct_names(x)

\S4method{ct_names}{Cycif}(x)

\S4method{ct_names}{CycifStack}(x)
}
\arguments{
\item{x}{A Cycif or CycifStack object.}
}
\value{
A character vector containing the names of cell types.
}
\description{
This function returns the names of cell types stored in a Cycif or CycifStack object.
}
