% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycifStack.R
\name{cyApply}
\alias{cyApply}
\alias{cyApply,CycifStack-method}
\title{Apply-like loop function for a CycifStack object}
\usage{
cyApply(x, ...)

\S4method{cyApply}{CycifStack}(x, fun, simplify = FALSE, as.CycifStack = TRUE, ...)
}
\arguments{
\item{x}{A CycifStack object}

\item{...}{Additional arguments passed to sapply/lapply functions.}

\item{fun}{function to apply to each Cycif object}

\item{simplify}{logical (FALSE by default). If TRUE, the function returns a vector or a matrix rather
than a list, if applicable. It uses 'sapply' instead of 'lapply' internally.}

\item{as.CycifStack}{logical (TRUE by default). If TRUE, the function attempts to convert the
output into a CycifStack object.}
}
\description{
Apply-like loop function for a CycifStack object
}
