% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-core.R, R/constructors-cycif.R,
%   R/methods-show.R
\docType{class}
\name{Cycif-class}
\alias{Cycif-class}
\alias{Cycif}
\alias{show,Cycif-method}
\title{Cycif Class}
\usage{
Cycif(
  ft_filename,
  path = ".",
  mask_type = c("cellRing", "cell"),
  mcmicro = FALSE,
  use_scimap = FALSE
)

\S4method{show}{Cycif}(object)
}
\arguments{
\item{ft_filename}{Character string, the name of the feature table file (a comma-separated file).}

\item{path}{Character string, the path to the feature table file. Default is ".".}

\item{mask_type}{Character vector specifying the type of mask used, options are "cellRing" or "cell".}

\item{mcmicro}{Logical, indicating whether the feature table comes from MCMICRO.}

\item{use_scimap}{Logical, indicating whether to use scimap python package for additional data processing. The default is FALSE (Note: scimap is used through reticulate but it's use is not fully implemented yet).}

\item{object}{a Cycif object}
}
\value{
A Cycif object containing information extracted from the feature table.
}
\description{
This class represents CyCIF data and its associated metadata.

The \code{Cycif} function creates a Cycif object from a given feature table file.
}
\details{
The \code{Cycif} class represents CyCIF data and its associated metadata, including sample name, file paths, imaging cycles, cell coordinates, cell properties, and more. This class is used to store and manipulate CyCIF data within R.

Instantiate and show a Cycif object

The \code{Cycif} function reads a feature table file and extracts information such as the number of cycles,
the number of cells, the list of antibodies, raw protein and DNA intensities, xy coordinates, and segment properties.
It then creates a Cycif object with the extracted information.
}
\section{Slots}{

\describe{
\item{\code{name}}{A character vector representing the sample name correspoding to the CyCIF dataset.}
\item{\code{file_paths}}{An object of class \code{file_paths} containing file paths for the data (not implemented).}
\item{\code{mask_type}}{A character vector specifying the type of mask used, e.g., 'cell', 'cellRing'.}
\item{\code{n_cycles}}{A numeric value indicating the total number of imaging cycles.}
\item{\code{n_cells}}{A numeric value representing the total number of cells in the dataset (including the ones out of ROIs and dropped from dnaFilter).}
\item{\code{abs_list}}{A data frame containing information about antibodies and the cycles they were used imaging cycles.}
\item{\code{dna}}{A data frame representing raw DNA data per cell with columns corresponding to imaging cycles.}
\item{\code{raw}}{A data frame representing raw protein data per cell with columns corresponding to imaging cycles.}
\item{\code{xy_coords}}{A data frame containing X and Y coordinates of cell centroids.}
\item{\code{segment_property}}{A data frame containing per-cell properties of image segmentations.}
\item{\code{used_cells}}{A numeric matrix indicating the used cells with values, where 0 or 1 corresponds to not available or available for each cycle. The number of columns corresponds to the number of cycles.}
\item{\code{rois}}{A list containing information about Regions of Interest (ROIs). Each element is an object of the roi class.}
\item{\code{within_rois}}{A logical vector specifying whether cells are within ROIs for each imaging cycle.}
\item{\code{dna_thres}}{A data frame representing DNA threshold values. The number of rows correspond to the number of cycles, and there are two columns that correspond to lower- and upper-boundary.}
\item{\code{log_normalized}}{A data frame representing log-transformed normalized data, which are computed by \code{normalize} function.}
\item{\code{logTh_normalized}}{A data frame representing log-transformed and thresholded normalized data, which are computed by \code{normalize} function.}
\item{\code{cell_types}}{A named list of \code{CellTypes} objects representing cell types.}
\item{\code{ld_coords}}{A list of \code{LDCoords} objects.}
\item{\code{cell_neighborhood}}{An object of class \code{CellNeighborhood} representing cell neighborhood analysis results.}
\item{\code{calls}}{A list of functions called.}
}
}

\seealso{
Other classes: \code{\link{CycifStack}}, \code{\linkS4class{CellTypes}}, \code{\linkS4class{LDCoords}}

\code{\link{Cycif}}, \code{\link{CycifStack}}, \code{\linkS4class{CellTypes}}
}
