% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{CycifStack-class}
\alias{CycifStack-class}
\title{A class that contains data of multiple samples from a CyCIF project.}
\description{
A class that contains data of multiple samples from a CyCIF project.
}
