% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CycifStack-utils.R
\name{nSamples}
\alias{nSamples}
\alias{names,CycifStack-method}
\alias{uniq_abs}
\alias{nCycles}
\alias{maxCycles}
\alias{nCells,CycifStack-method}
\alias{length,CycifStack-method}
\title{Accessing slots in a CellTypeCycifStack object}
\usage{
nSamples(x)

\S4method{names}{CycifStack}(x)

uniq_abs(x)

nCycles(x)

maxCycles(x)

\S4method{nCells}{CycifStack}(x)

\S4method{length}{CycifStack}(x)
}
\arguments{
\item{x}{A CycifStack object}
}
\description{
Accessing slots in a CellTypeCycifStack object
}
