% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-subset.R
\name{[,CycifStack,ANY,ANY,ANY-method}
\alias{[,CycifStack,ANY,ANY,ANY-method}
\alias{[,CycifStack-method}
\alias{[[,CycifStack,ANY,ANY-method}
\alias{[<-,CycifStack,ANY,ANY,ANY-method}
\alias{[<-,CycifStack-method}
\alias{[[<-,CycifStack,ANY,ANY,ANY-method}
\alias{[[<-,CycifStack,ANY,ANY-method}
\title{Subsetting samples from a CycifStack object}
\usage{
\S4method{[}{CycifStack,ANY,ANY,ANY}(x, i = NULL)

\S4method{[[}{CycifStack,ANY,ANY}(x, i = NULL)

\S4method{[}{CycifStack,ANY,ANY,ANY}(x, i) <- value

\S4method{[[}{CycifStack,ANY,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{A CycifStack object}

\item{i}{A numeric or character vector. If numeric, the values are integers
representing indices for the Cycif objects. If character, names of the Cycif
objects.}

\item{value}{A Cycif object to be inserted or replaced.}
}
\description{
Subsetting samples from a CycifStack object
}
