% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis.R
\name{defineTumorBed}
\alias{defineTumorBed}
\title{Identify Tumor bed based on dbscan clusters}
\usage{
defineTumorBed(x, ...)

defineTumorBed(x, ...)
}
\arguments{
\item{x}{A Cycif obj}

\item{strict}{logical. if cell_types should be strict.}

\item{dth}{radius for dbscan (eps)}

\item{min.pts}{minimum points for dbscan}

\item{parallel}{logical. The default is TRUE.}

\item{n.cores}{number of cores.}
}
\description{
Identify Tumor bed based on dbscan clusters
}
