% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-tumor-border.R
\name{defineTumorBorder}
\alias{defineTumorBorder}
\alias{defineTumorBorder,Cycif-method}
\alias{defineTumorBorder,CycifStack-method}
\title{Define tumor bed regions within each Cycif dataset.}
\usage{
defineTumorBorder(x, ...)

\S4method{defineTumorBorder}{Cycif}(
  x,
  strict = FALSE,
  dth,
  parallel = TRUE,
  min.pts = 3,
  n.cores = 1,
  ct_name = "default",
  verbose = TRUE,
  cancer.cts = "Tumor"
)

\S4method{defineTumorBorder}{CycifStack}(
  x,
  strict = FALSE,
  dth,
  parallel = TRUE,
  min.pts = 3,
  n.cores = 1,
  ct_name = "default",
  verbose = TRUE,
  cancer.cts = "Tumor"
)
}
\arguments{
\item{x}{A Cycif or CycifStack object.}

\item{...}{Additional arguments (unused).}

\item{strict}{Logical, specifying whether to perform strict cell type calling (see \code{defineCellTypes}).}

\item{dth}{Numeric, the distance threshold for identifying adjacent cells.}

\item{parallel}{Logical, indicating whether to use parallel processing for edge point calculation in \code{concaveman}.}

\item{min.pts}{Integer, the minimum number of points required to form a cluster.}

\item{n.cores}{Integer, the number of CPU cores to use for parallel processing.}

\item{ct_name}{Character, the name of the cell type column (used in \code{cell_types}).}

\item{verbose}{Logical, indicating whether to display progress messages.}

\item{cancer.cts}{Character vector of cell type(s) treated as tumor cells. Default "Tumor".}
}
\value{
A list containing information about tumor bed regions, including distance threshold, cell coordinates, regions within tumor bed, edge points, cell types, clusters, and neighbors.
}
\description{
This function identifies tumor bed regions within each Cycif dataset and assigns cells to these regions based on their spatial coordinates and cell types.
}
\details{
To define tumor bed regions within CyCIF data, the following steps are performed:
\enumerate{
\item If the \code{dth} parameter is not provided, it calculates an optimal distance threshold based on the MajorAxisLength of cells.
\item Adjusts the Y-coordinate values to ensure proper orientation.
\item Identifies cell types within the dataset, allowing for strict or non-strict matching.
\item Separates tumor cells from non-tumor cells based on their cell types.
\item Performs DBSCAN clustering on tumor cells to identify clusters of adjacent cells.
\item Determines neighbors using NN (Fast Regular Nearest Neighbors) only for tumor cells.
\item Calculates edge points for the tumor clusters using the concaveman algorithm.
\item Determines whether cells are within the tumor bed regions based on edge point polygons.
\item Returns a list containing various information about the tumor bed regions and assigned cells.
}
}
\seealso{
\code{\link{cyApply}} \code{\link{defineCellTypes}} \code{\link{concaveman}}
}
