% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-tumor-stroma.R
\name{defineTumorStroma}
\alias{defineTumorStroma}
\alias{defineTumorStroma,Cycif-method}
\title{Define tumor bed regions within each Cycif dataset.}
\usage{
defineTumorStroma(x, ...)

\S4method{defineTumorStroma}{Cycif}(
  x,
  n.cores = 7,
  n.cells.per.seg = 100,
  n.cells.per.tumor.core = 10,
  cancer.cts = c("Cancer"),
  ct_name = "default",
  dth,
  concavity = 0.8,
  plot = FALSE,
  ...
)
}
\arguments{
\item{x}{A CyCIF object.}

\item{...}{Additional arguments (unused).}

\item{n.cores}{Integer, the number of CPU cores to use for parallel processing.}

\item{n.cells.per.seg}{Integer, the minimum number of cells per segment.}

\item{n.cells.per.tumor.core}{Integer, the minimum number of cells per tumor region.}

\item{cancer.cts}{Character vector of cell type(s) treated as tumor cells. Default "Cancer".}

\item{ct_name}{Character, the name of the cell type column (used in \code{cell_types}).}

\item{dth}{Numeric, the distance threshold for identifying adjacent cells.}

\item{concavity}{a relative measure of concavity. 1 results in a relatively detailed shape, Infinity results in a convex hull. You can use values lower than 1, but they can produce pretty crazy shapes (\code{concaveman}).}

\item{plot}{Logical, indicating whether to display plots.}
}
\value{
A list containing information about tumor bed regions, including distance threshold, cell coordinates, regions within tumor bed, edge points, cell types, clusters, and neighbors.
}
\description{
This function identifies tumor bed regions within each Cycif dataset and assigns cells to these regions based on their spatial coordinates and cell types.
}
\details{
The \code{dist2tumorBorder} function calculates the distance of each cell in the CyCIF dataset to the nearest tumor border. It performs the following steps:
\enumerate{
\item It extracts the coordinates of cells from the CyCIF data, converting them into an sf object.
\item It first identifies all the cells within ROIs and compute the boundary for each tissue segment. Note that a segment containing fewer than \code{n.cells.per.seg} cells is ignored.
\item For each segment, it identifies tumor cells and their clusters using DBSCAN clustering based on the specified distance threshold (\code{dth}).
\item It computes the tumor border polygons by using the DBSCAN and concaveman algorithm on each tumor cell cluster.
\item It determines if each non-tumor cell is inside or outside of the tumor border polygons using the \code{st_distance} function. For the cells inside any polygons are assigned to the tumor bed, and their distance to the nearest border is calculated (expressed as negative value)
\item The function returns a list containing the calculated distances and the tumor border polygons.
}
}
\seealso{
\code{\link{concaveman}}, \code{\link{dbscan}}, \code{\link{frNN}}, \code{\link{st_distance}}, \code{\link{st_intersects}}, \code{\link{st_union}}, \code{\link{st_boundary}}
}
