% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis.R
\name{dist2tumBorder}
\alias{dist2tumBorder}
\title{#'}
\usage{
dist2tumBorder(x, n.cores = 7, dth)
}
\arguments{
\item{x}{A Cycif obj}

\item{n.cores}{number of cores.}

\item{dth}{radius for dbscan (eps)}

\item{min.pts}{minimum points for dbscan}
}
\description{
#'
}
