% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-distance.R
\name{dist2tumorBorder}
\alias{dist2tumorBorder}
\alias{dist2tumorBorder,Cycif-method}
\title{Calculate the Distance to Tumor Border in CyCIF data.}
\usage{
dist2tumorBorder(x, ...)

\S4method{dist2tumorBorder}{Cycif}(
  x,
  n.cores = 7,
  dth,
  th,
  minPts = 3,
  concavity = 0.8,
  plot = FALSE,
  ...
)
}
\arguments{
\item{x}{A CyCIF object.}

\item{...}{Additional arguments to be passed.}

\item{n.cores}{Integer, number of CPU cores to use for parallel processing.}

\item{dth}{Numeric, the distance threshold used for tumor border detection.}

\item{th}{Numeric, distance threshold used when assigning cells to the border/interior split.}

\item{minPts}{Integer, the minimum number of points required to form a cluster in DBSCAN.}

\item{concavity}{a relative measure of concavity. 1 results in a relatively detailed shape, Infinity results in a convex hull. You can use values lower than 1, but they can produce pretty crazy shapes (\code{concaveman}).}

\item{plot}{Logical, whether to plot the tumor border polygons. Default FALSE.}
}
\value{
A list containing the calculated distances and tumor border polygons.
}
\description{
This function calculates the distance of each cell in CyCIF data to the nearest tumor border. The tumor border is defined based on the spatial arrangement of tumor cells within the dataset.
}
