% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-distance.R
\name{distAdjacentCells}
\alias{distAdjacentCells}
\alias{distAdjacentCells,Cycif-method}
\alias{distAdjacentCells,CycifStack-method}
\title{Calculate the Distance Between Adjacent Cells in CyCIF Data.}
\usage{
distAdjacentCells(x, ...)

\S4method{distAdjacentCells}{Cycif}(x)

\S4method{distAdjacentCells}{CycifStack}(x)
}
\arguments{
\item{x}{A Cycif or CycifStack object.}

\item{...}{Additional arguments (unused).}
}
\value{
A numeric distance threshold (mean cell diameter + 2 SD, in pixels).
}
\description{
Calculate the Distance Between Adjacent Cells in CyCIF Data.
}
