% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-basic.R
\name{dna_thres}
\alias{dna_thres}
\alias{dna_thres,Cycif-method}
\title{Get the DNA threshold in a Cycif object}
\usage{
dna_thres(x)

\S4method{dna_thres}{Cycif}(x)
}
\arguments{
\item{x}{A Cycif object.}
}
\value{
A data frame containing the DNA threshold in the Cycif object.
}
\description{
The \code{dna_thres} function returns the DNA threshold in a Cycif object.
}
\seealso{
\code{\link{nSamples}}: Get the number of samples in a CycifStack object.
\code{\link{names}}: Get the names of samples in a CycifStack object.
\code{\link{maxCycles}}: Get the maximum number of cycles in a CycifStack object.
\code{\link{within_rois}}: Check if samples in a CycifStack are within ROIs.
\code{\link{length}}: Equivalent to \code{nSamples} for CycifStack objects.
}
