% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCellTypes.R
\name{findCellTypes}
\alias{findCellTypes}
\title{Perform a cell type calling function and set cell types in a Cycif or CycifStack object}
\usage{
findCellTypes(x, ...)
}
\description{
Perform a cell type calling function and set cell types in a Cycif or CycifStack object
}
