% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{FindClusters}
\alias{FindClusters}
\alias{FindClusters,matrix-method}
\alias{FindClusters,Cycif-method}
\alias{FindClusters,CycifStack-method}
\title{Find clusters using Seurat functions}
\usage{
FindClusters(x, ...)

\S4method{FindClusters}{matrix}(
  x,
  k.param = 20,
  initial.membership = NULL,
  node.sizes = NULL,
  resolution = 0.8,
  algorithm = 1,
  with.labels = FALSE,
  ...
)

\S4method{FindClusters}{Cycif}(
  x,
  ld_name,
  k.param = 20,
  initial.membership = NULL,
  node.sizes = NULL,
  resolution = 0.8,
  algorithm = 1,
  ...
)

\S4method{FindClusters}{CycifStack}(
  x,
  ld_name,
  k.param = 20,
  initial.membership,
  node.sizes,
  resolution = 0.8,
  algorithm = 1,
  ...
)
}
\arguments{
\item{x}{Object of class "CycifStack"}

\item{...}{Additional arguments to be passed}

\item{k.param}{Integer, number of nearest neighbors}

\item{initial.membership}{Vector of initial cluster assignments}

\item{node.sizes}{List of node sizes for cluster centers}

\item{resolution}{Double, resolution of the clustering algorithm}

\item{algorithm}{Integer, clustering algorithm to use}

\item{with.labels}{Logical scalar. Whether to include labels.}

\item{ld_name}{Name of the label to use for clustering}
}
\value{
an object of the same class with x

An integer vector with cluster assignments for each cell.

A modified \code{\link{Cycif}} object with cluster assignments added to the specified layout.

The input CycifStack object with the cluster assignments added
}
\description{
This function finds clusters in a given matrix using Seurat::FindNeighbors and Seurat::FindClusters functions.
}
\examples{
\dontrun{
mat <- matrix(rnorm(100), nrow = 10)
FindClusters(mat, k.param = 5)
}

\dontrun{
data("exampleCycif")
clusters <- FindClusters(exampleCycif, "Layout1", k.param = 10)
}

\dontrun{
FindClusters(x, ld_name = "label_name")
}
}
