% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{frNN-class}
\alias{frNN-class}
\title{Class "frNN" - output of dbscan::frNN()}
\description{
Class "frNN" - output of dbscan::frNN()
}
\section{Slots}{

\describe{
\item{\code{id}}{A list of integer vector (length of the number of cell neighborhoods). Each vector contains the ids of the fixed radius nearest neighbors.}

\item{\code{dist}}{A list with distances (same structure as id)}

\item{\code{eps}}{A numeric vector (scalar) containing neighborhood radius eps that was used.}

\item{\code{sort}}{A logical value indicating whether the distances are sorted.}
}}

\seealso{
\code{\link{dbscan::frNN}}
}
