% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{gates}
\alias{gates}
\title{Show gates in a Cycif/CycifStack/CellTypeCycif/CellTypeCicyfStack object}
\usage{
gates(x)
}
\description{
Show gates in a Cycif/CycifStack/CellTypeCycif/CellTypeCicyfStack object
}
