% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-qc.R
\name{hist_1d}
\alias{hist_1d}
\title{1D Histogram Plot}
\usage{
hist_1d(x, n = 1000, ths, mar = c(3, 4, 4, 2) + 0.1, brks1, ttl1)
}
\arguments{
\item{x}{A numeric vector for which the histogram is to be plotted.}

\item{n}{The number of bins or breaks for the histogram (default is 1000).}

\item{ths}{A vector of threshold values to mark on the plot (default is NULL).}

\item{mar}{A numerical vector of length 4 specifying the margin sizes (default is c(3, 4, 4, 2) + 0.1).}

\item{brks1}{A vector of pre-defined breaks for the histogram (default is NULL).}

\item{ttl1}{The title for the histogram plot (default is NULL).}
}
\description{
This function creates a 1D histogram plot for a numeric vector.
}
\details{
\itemize{
\item The \code{hist_1d} function creates a 1D histogram plot to visualize the distribution of a numeric vector.
\item You can specify the number of bins with the \code{n} parameter or provide custom break points with the \code{brks1} parameter.
\item Threshold values can be added to the plot using the \code{ths} parameter.
\item Additional customization of the plot can be achieved using various graphical parameters.
}
}
