% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{LdClustering}
\alias{LdClustering}
\alias{LdClustering,matrix-method}
\alias{LdClustering,Cycif-method}
\alias{LdClustering,CycifStack-method}
\title{Find clusters using Seurat functions}
\usage{
LdClustering(x, ...)

\S4method{LdClustering}{matrix}(
  x,
  k.param = 20,
  initial.membership = NULL,
  node.sizes = NULL,
  resolution = 0.8,
  algorithm = 1,
  with.labels = FALSE,
  method = "igraph",
  ...
)

\S4method{LdClustering}{Cycif}(
  x,
  ld_name,
  k.param = 20,
  initial.membership,
  node.sizes,
  resolution = 0.8,
  algorithm = 1,
  method = "igraph",
  ...
)

\S4method{LdClustering}{CycifStack}(
  x,
  ld_name,
  k.param = 20,
  initial.membership,
  node.sizes,
  resolution = 0.8,
  algorithm = 1,
  method = "igraph",
  ...
)
}
\arguments{
\item{x}{An matrix, Cycif, or CycifStack object containing data to cluster.}

\item{...}{Additional arguments passed to the clustering function.}

\item{k.param}{An integer specifying the number of nearest neighbors.}

\item{initial.membership}{A vector specifying the initial cluster membership of each cell.}

\item{node.sizes}{A vector specifying the size of each node in the clustering tree.}

\item{resolution}{A numeric scalar specifying the granularity of the clustering.}

\item{algorithm}{An integer specifying the clustering algorithm to use.}

\item{with.labels}{Logical scalar indicating whether to include labels.}

\item{method}{Community-detection method passed to \code{Seurat::FindClusters}. Default "igraph".}

\item{ld_name}{A character scalar indicating the name of the layout to use.}
}
\value{
A modified object with cluster assignments added to the specified layout (for a \code{matrix} object).
An object of the same class as \code{x} with cluster assignments added (for a \code{Cycif} or \code{CycifStack} object).
}
\description{
Find clusters using Seurat functions
}
