% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{LdRunUMAP}
\alias{LdRunUMAP}
\alias{LdRunUMAP,Cycif-method}
\alias{LdRunUMAP,CycifStack-method}
\title{Run visualization with dimensionality reduction (t-SNE or U-MAP) on CyCIF data.}
\usage{
LdRunUMAP(x, ...)

\S4method{LdRunUMAP}{Cycif}(
  x,
  norm_type = c("raw", "log", "logTh"),
  ld_name,
  ct_name = "default",
  ncells.per.smpl = NULL,
  used.abs,
  used.cts,
  strict = TRUE,
  n_neighbors = 20,
  smpl.seed = 123,
  umap.seed = 12345,
  ...
)

\S4method{LdRunUMAP}{CycifStack}(
  x,
  norm_type = c("raw", "log", "logTh"),
  ld_name,
  ct_name = "default",
  ncells.per.smpl = NULL,
  used.abs,
  used.cts,
  strict = TRUE,
  n_neighbors = 20,
  smpl.seed = 123,
  umap.seed = 12345,
  ...
)
}
\arguments{
\item{x}{A Cycif or CycifStack object.}

\item{...}{Additional arguments passed to uwot::umap().}

\item{norm_type}{Character string specifying the type of expression values to use for dimensionality reduction. It should be one of "raw," "log," or "logTh" indicating whether to use raw or normalized data.}

\item{ld_name}{Character string specifying the name to assign to the dimensionality reduction results. This name will be used to retrieve the data later.}

\item{ct_name}{Character string specifying the cell type name to be used for analysis. Default is "default."}

\item{ncells.per.smpl}{Numeric value specifying the number of cells per sample. If NULL, the maximum number of cells per sample will be used.}

\item{used.abs}{Character vector containing a set of antibodies used for UMAP computation.}

\item{used.cts}{Character vector containing a set of cell types used for UMAP computation.}

\item{strict}{Logical indicating whether strict filtering of cell types should be applied. Default is TRUE.}

\item{n_neighbors}{Numeric specifying the number of neighbors to consider during dimensionality reduction.}

\item{smpl.seed}{Numeric specifying the random seed used for per-sample cell subsampling. Default 123.}

\item{umap.seed}{Numeric specifying the initial seed for reproducible UMAP results.}
}
\value{
A Cycif or CycifStack object with added UMAP or t-SNE coordinates and related metadata.
}
\description{
This function performs dimensionality reduction (t-SNE or U-MAP) on CyCIF data, either for a single sample or across multiple samples (CycifStack). It generates UMAP or t-SNE coordinates for visualization purposes.
}
\details{
The \code{LdRunUMAP} function performs dimensionality reduction on CyCIF data using UMAP or t-SNE algorithms, depending on the specified parameters. It generates UMAP or t-SNE coordinates for visualization purposes and adds them to the provided Cycif or CycifStack object. The resulting object contains the computed coordinates, along with metadata such as the type of dimensionality reduction, normalization type, used antibodies, and used cell types.
}
\seealso{
\code{\linkS4class{LDCoords}}, \code{\link{Cycif}}, \code{\link{CycifStack}}
}
