% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion-list.R
\name{list2Cycif}
\alias{list2Cycif}
\alias{list2Cycif,list-method}
\title{Converting a list to a Cycif object}
\usage{
list2Cycif(x)

\S4method{list2Cycif}{list}(x)
}
\arguments{
\item{x}{A list object}
}
\value{
A Cycif object
}
\description{
Converting a list to a Cycif object
}
\seealso{
\code{\link{list2CycifStack}}, \code{\link{Cycif2CycifStack}},
\code{\link{Cycif-class}}, \code{\link{CycifStack-class}}
}
