% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion-stack.R
\name{list2CycifStack}
\alias{list2CycifStack}
\alias{list2CycifStack,list-method}
\alias{Cycif2CycifStack}
\alias{Cycif2CycifStack,Cycif-method}
\title{Convert a list of Cycif or CycifStack objects to a CycifStack object}
\usage{
list2CycifStack(x)

\S4method{list2CycifStack}{list}(x)

Cycif2CycifStack(x)

\S4method{Cycif2CycifStack}{Cycif}(x)
}
\arguments{
\item{x}{A list containing Cycif or CycifStack objects.}
}
\value{
A CycifStack object containing information from the input list of Cycif or CycifStack objects.
}
\description{
The \code{list2CycifStack} and \code{Cycif2CycifStack} functions convert a list of Cycif or CycifStack objects into a single CycifStack object.
}
\details{
The \code{list2CycifStack} function takes a list of Cycif or CycifStack objects and combines them into a single CycifStack.
It extracts information such as the number of samples, names, antibodies, cycles, and cell counts from each input object,
and creates a new CycifStack object in which each element is a Cycif object.
The \code{Cycif2CycifStack} function handles a special case where a single Cycif object is converted to a CycifStack object.
}
\seealso{
\code{\link{list2Cycif}}, \code{\link{list2CycifStack}}, \code{\link{Cycif2CycifStack}}, \code{\link{Cycif}}, \code{\link{CycifStack}}, \code{\linkS4class{CellTypes}}
}
