% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-basic.R, R/setters-core.R
\name{names,Cycif-method}
\alias{names,Cycif-method}
\alias{names,CycifStack-method}
\alias{names<-,CycifStack-method}
\title{Get the name(s) of a Cycif or CycifStack object}
\usage{
\S4method{names}{Cycif}(x)

\S4method{names}{CycifStack}(x)

\S4method{names}{CycifStack}(x) <- value
}
\arguments{
\item{x}{A Cycif or CycifStack object.}

\item{value}{Character vector of new sample names.}
}
\value{
For a Cycif object, its sample name. For a CycifStack object, the
names of all samples it contains.
}
\description{
Get the name(s) of a Cycif or CycifStack object
}
