% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-dimensions.R
\name{nCells}
\alias{nCells}
\alias{nCells,Cycif-method}
\alias{nCells,CycifStack-method}
\title{Get the number of cells in each sample of a CycifStack object}
\usage{
nCells(x)

\S4method{nCells}{Cycif}(x)

\S4method{nCells}{CycifStack}(x)
}
\arguments{
\item{x}{A CycifStack object.}
}
\value{
A numeric vector containing the number of cells in each sample of the CycifStack object.
}
\description{
The \code{nCells} function returns the number of cells in each sample of a CycifStack object.
Note that the cells are not assessed for their QC status. For the number of cells dropped
through DNAFiter or ROI selection, use other methods.
}
\seealso{
\code{\link{nSamples}}: Get the number of samples in a CycifStack object.
\code{\link{names}}: Get the names of samples in a CycifStack object.
\code{\link{maxCycles}}: Get the maximum number of cycles in a CycifStack object.
\code{\link{within_rois}}: Check if samples in a CycifStack are within ROIs.
\code{\link{length}}: Equivalent to \code{nSamples} for CycifStack objects.
}
