% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycif.R
\name{nCycles<-}
\alias{nCycles<-}
\title{Set nCycles - note this should run before cell type calling}
\usage{
nCycles(x, ...) <- value
}
\description{
Set nCycles - note this should run before cell type calling
}
