% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-dimensions.R, R/setters-core.R
\name{nCycles}
\alias{nCycles}
\alias{nCycles,Cycif-method}
\alias{nCycles,CycifStack-method}
\alias{nCycles<-}
\alias{nCycles<-,Cycif-method}
\alias{nCycles<-,CycifStack-method}
\title{Get or Set the number of cycles in a CycifStack object}
\usage{
nCycles(x)

\S4method{nCycles}{Cycif}(x)

\S4method{nCycles}{CycifStack}(x)

nCycles(x, ...) <- value

\S4method{nCycles}{Cycif}(x) <- value

\S4method{nCycles}{CycifStack}(x) <- value
}
\arguments{
\item{x}{A CycifStack object.}

\item{...}{Additional arguments (unused; for \verb{nCycles<-}).}

\item{value}{A numeric scalar indicating the desired number of cycles.}
}
\value{
The \code{nCycles} function returns a numeric vector containing the number of cells in each sample of the CycifStack object.
The \verb{nCycles<-} function returns a modified CycifStack object with the updated number of cycles.
}
\description{
The \code{nCycles} function returns the number of cells in each sample of a CycifStack object.
The \verb{nCycles<-} function sets the number of cycles in a CycifStack object. It trims samples that do not reach the specified cycle value.
}
\details{
This \code{nCycles} function calculates the number of cells in each sample of a CycifStack object and returns the results as a numeric vector.
The \verb{nCycles<-} function sets the number of cycles in a CycifStack object to the specified value. It trims samples that do not reach the specified cycle value, ensuring that all samples have the same number of cycles.
}
\note{
This function is designed to accept a numeric scalar (\code{value}) as the desired number of cycles. In the future, it will not accept a vector, but only a scalar value.
}
\seealso{
\code{\link{nSamples}}: Get the number of samples in a CycifStack object.
\code{\link{names}}: Get the names of samples in a CycifStack object.
\code{\link{maxCycles}}: Get the maximum number of cycles in a CycifStack object.
\code{\link{within_rois}}: Check if samples in a CycifStack are within ROIs.
\code{\link{length}}: Equivalent to \code{nSamples} for CycifStack objects.
}
