% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-spatial.R
\docType{class}
\name{NN-class}
\alias{NN-class}
\title{Class "NN" - output of dbscan::frNN() or dbscan::kNN()}
\description{
Class "NN" - output of dbscan::frNN() or dbscan::kNN()
}
\section{Slots}{

\describe{
\item{\code{type}}{A character vector specifying the type of nearest neighbors. Possible values are 'frnn' or 'knn'.}

\item{\code{id}}{A list of integer vector (length of the number of cell neighborhoods). Each vector contains the ids of the fixed radius nearest neighbors.}

\item{\code{dist}}{A list with distances (same structure as id)}

\item{\code{k}}{A numeric vector (scalar) containing the number of neighbors k that was used.}

\item{\code{eps}}{A numeric vector (scalar) containing neighborhood radius eps that was used.}

\item{\code{sort}}{A logical value indicating whether the distances are sorted.}
}}

\seealso{
\code{\link[dbscan]{frNN}} \code{\link[dbscan]{kNN}}
}
