% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pData.R
\name{pData<-}
\alias{pData<-}
\title{Insert metadata into a Cycif object}
\usage{
pData(x, ...) <- value
}
\arguments{
\item{x}{A Cycif object}

\item{value}{input value of a data frame object}

\item{by}{the column name in the value. The variable should contain the sample names, which can be assessed by \code{names(x)}. The column name should be set as The variable is set as the 'id'.}
}
\description{
Insert metadata into a Cycif object
}
