% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{roiFilter}
\alias{roiFilter}
\alias{roiFilter,Cycif-method}
\title{Set ROIs in a Cycif object interactively by clicking on the plot (discontinued)}
\usage{
roiFilter(x, ...)

\S4method{roiFilter}{Cycif}(x, rois)
}
\arguments{
\item{x}{A Cycif object.}

\item{...}{Additional arguments (unused).}

\item{rois}{A list of ROI definitions to set (see \code{\link{importROIs}}).}
}
\value{
A Cycif object with the defined ROIs.
}
\description{
This function allows interactive setting of regions of interest (ROIs) in a Cycif object. ROIs can be defined as either positive or negative based on their intended use.
}
\details{
This function provides an interactive way to define ROIs within a Cycif object. ROIs can be classified as either "positive" or "negative" based on their intended use. Users can set these ROIs by selecting points on the plot. For positive ROIs, cells within the polygonal boundaries are considered within the ROI. For negative ROIs, cells outside the polygonal boundaries are considered within the ROI. Users can set multiple ROIs of each type by following the prompts.
}
\seealso{
\code{\link{importROIs}}, \code{\link{setWithinROIs}}
}
