% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-basic.R
\name{segProp}
\alias{segProp}
\alias{segProp,Cycif-method}
\title{Get the segmentation property in a Cycif object}
\usage{
segProp(x)

\S4method{segProp}{Cycif}(x)
}
\arguments{
\item{x}{A Cycif object.}
}
\value{
A character vector containing the segmentation property in the Cycif object.
}
\description{
The \code{segProp} function returns the segmentation property in a Cycif object.
}
\seealso{
\code{\link{nSamples}}: Get the number of samples in a CycifStack object.
\code{\link{names}}: Get the names of samples in a CycifStack object.
\code{\link{maxCycles}}: Get the maximum number of cycles in a CycifStack object.
\code{\link{within_rois}}: Check if samples in a CycifStack are within ROIs.
\code{\link{length}}: Equivalent to \code{nSamples} for CycifStack objects.
}
