% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-cell-neighborhood.R
\name{setDist}
\alias{setDist}
\alias{setDist,CellNeighborhood-method}
\title{set distance to tumorBorder for NN objects}
\usage{
setDist(x, ...)

\S4method{setDist}{CellNeighborhood}(x, value)
}
\arguments{
\item{x}{A NN object.}

\item{...}{Additional arguments (unused).}

\item{value}{A numeric vector specifying the distance to tumor border for each cell.}
}
\description{
set distance to tumorBorder for NN objects
}
