% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{setWithinROIs}
\alias{setWithinROIs}
\alias{setWithinROIs,Cycif-method}
\title{Set within_roi slot in a Cycif object}
\usage{
setWithinROIs(x, ...)

\S4method{setWithinROIs}{Cycif}(x)
}
\arguments{
\item{x}{A Cycif object.}

\item{...}{Additional arguments (unused).}
}
\value{
A Cycif object with cells marked as within or outside ROIs in within_rois slot
}
\description{
Compute if each cell is within any positive ROIs and outside any negative ROIs. The value is TRUE if that is the case, FALSE otherwise.
This should be used after importROIs functions are called so ROIs are set in each Cycif object.
}
\details{
This function determines which cells in a Cycif object are within the defined ROIs. It uses information about the direction (positive or negative), cycle, and coordinates of the ROIs to classify cells as either within or outside ROIs. Cells are marked as within ROIs if they meet the following conditions:
\itemize{
\item They are within the polygonal boundaries of positive ROIs.
\item They are outside the polygonal boundaries of negative ROIs.
}
}
