% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycif.R
\name{statUsedCells,Cycif-method}
\alias{statUsedCells,Cycif-method}
\title{Summarize available cells at each cycle}
\usage{
statUsedCells(x, cumulative=TRUE, ratio=TRUE, use_rois=TRUE)
}
\arguments{
\item{cumulative}{logical. If TRUE, return the number of cells available from cycle 1 to N. FALSE, return the number of cells in each cycle. Default is TRUE.}

\item{ratio}{logical. If TRUE, return the ratio of available cells over all the cells. Default is TRUE.}

\item{use_rois}{logical. If TRUE, compute available cells within pre-specified ROIs.}
}
\description{
Summarize available cells at each cycle
}
