% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-cell-neighborhood.R
\name{tcnClust}
\alias{tcnClust}
\alias{tcnClust,data.frame-method}
\title{Cluster and Sort Recurrent Cell Neighbors (RCN) for CyCIF or CyCIFStack Objects}
\usage{
tcnClust(nn, ...)

\S4method{tcnClust}{data.frame}(
  nn,
  g = 50,
  seed = 123,
  sort.by = "CD8T",
  sort.type = c("freq", "count"),
  sort.smpls = c("all", "selected"),
  data.type = c("ct_exp", "ct"),
  extrapolate = FALSE,
  mc.cores = 1
)
}
\arguments{
\item{nn}{An object containing RCN information, typically obtained from 'computeCN'.}

\item{...}{Additional arguments (unused).}

\item{g}{The number of clusters to create.}

\item{seed}{The random seed for reproducibility.}

\item{sort.by}{The cell type to sort clusters by (e.g., "CD8T").}

\item{sort.type}{The type of data to use for sorting: "freq" (relative frequencies) or "count" (counts).}

\item{sort.smpls}{The subset of samples to use for sorting: "all" (entire dataset) or "selected" (selected cells).}

\item{data.type}{The type of data to use for clustering: "ct_exp" (cell type and expression data) or "ct" (cell type data only).}

\item{extrapolate}{Whether to extrapolate clusters to the entire dataset (TRUE) or not (FALSE).}

\item{mc.cores}{The number of CPU cores to use for parallel processing.}
}
\value{
An updated 'nn' object with clustering and sorting information.
}
\description{
This function clusters and sorts the Recurrent Cell Neighbors (RCN) for CyCIF or CyCIFStack objects.
It clusters cells based on their RCN profiles, sorts clusters based on the specified cell type,
and optionally extrapolates the clustering to the entire dataset.
}
\details{
The \code{tcnClust} function uses the provided \code{nn} object to perform clustering and classification of cells based on their neighborhood relationships. It allows you to specify the number of clusters (\code{g}), the cell type to sort clusters by (\code{sort.by}), and other clustering parameters.
The clustering process results in the classification of cells into distinct clusters, and the function provides information about these clusters, including the mean frequencies, counts, and more.
By specifying different options for \code{sort.by}, \code{sort.type}, and \code{sort.smpls}, you can customize the sorting behavior of clusters based on cell types and data types.
Additionally, you can choose to extrapolate clusters to the entire dataset using the \code{extrapolate} argument, which can be helpful for analyzing the overall dataset.
}
\seealso{
\code{\link{computeCN}}
}
