% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{threshold}
\alias{threshold}
\alias{threshold<-}
\alias{threshold<-,CycifStack-method}
\alias{threshold<-,CellTypeCycifStack-method}
\title{Show or set a threshold in a Cycif or CycifStack object}
\usage{
threshold(x)

threshold(x, ...) <- value

\S4method{threshold}{CycifStack}(x, strict = FALSE, ...) <- value

\S4method{threshold}{CellTypeCycifStack}(x, ...) <- value
}
\description{
Show or set a threshold in a Cycif or CycifStack object
}
