% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{transform}
\alias{transform}
\title{Transform Numeric Data}
\usage{
transform(
  r,
  method = c("log", "logTh", "Th", "invlog"),
  th,
  p_thres = 0.5,
  trim = 0.001
)
}
\arguments{
\item{r}{A numeric vector or array.}

\item{method}{The transformation method to apply. Choose from "log", "logTh", "Th", or "invlog".
\itemize{
\item "log": Log-transforms the data using log1p().
\item "logTh": Log-transforms the data and trims outliers.
\item "Th": Trims outliers.
\item "invlog": Reverts the data back to its original scale.
}}

\item{th}{The threshold value for transformation (only for "Th" and "logTh" methods).}

\item{p_thres}{The threshold value for data scaling (only for "logTh" and "Th" methods).}

\item{trim}{The quantile-based trimming level for outlier removal (default is 1e-3).}
}
\value{
Transformed numeric data.
}
\description{
This function applies various transformations to numeric data based on the specified method.
}
\seealso{
\code{\link{normalize}}
}
