% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlnPlot.R
\name{vlnPlot,CycifStack-method}
\alias{vlnPlot,CycifStack-method}
\title{#' @export
setMethod("vlnPlot", "CycifStack",
function(x,ab,type=c("raw","normalized"),...){
require(tidyr)
require(ggplot2)}
\usage{
\S4method{vlnPlot}{CycifStack}(
  x,
  type = c("raw", "normalized"),
  xaxis = c("celltype", "smpl"),
  ab = "PDL1",
  cts,
  ctype = "all",
  sample = "all",
  ttl
)
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{if(missing(type))\{
  type <- "normalized"
\}
}\if{html}{\out{</div>}}
}
\details{

}
