% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-basic.R
\name{within_rois}
\alias{within_rois}
\alias{within_rois,Cycif-method}
\title{Check if samples in a CycifStack are within ROIs}
\usage{
within_rois(x)

\S4method{within_rois}{Cycif}(x)
}
\arguments{
\item{x}{A CycifStack object.}
}
\value{
A logical vector indicating whether each sample is within an ROI.
}
\description{
The \code{within_rois} function checks if samples in a CycifStack object are within regions of interest (ROIs).
}
\seealso{
\code{\link{nSamples}}: Get the number of samples in a CycifStack object.
\code{\link{names}}: Get the names of samples in a CycifStack object.
\code{\link{maxCycles}}: Get the maximum number of cycles in a CycifStack object.
}
