% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{AbsSummary}
\alias{AbsSummary}
\title{A graphical summary of cycles and antibodies of a CycifStack class.}
\usage{
AbsSummary(x, show.cycles.in.row = FALSE, ...)
}
\arguments{
\item{x}{A CycifStack object.}

\item{show.cycles.in.row}{Logical, indicating whether to display cycle information in row labels.
FALSE by default. If TRUE, the output plot shows the number of cycles each sample was stained for.}

\item{...}{Additional parameters to be passed to the \code{graphics::image} function.}
}
\value{
The function generates a graphical absolute summary plot and returns \code{NULL}.
}
\description{
Generates an absolute summary plot for a CycifStack object.
}
\details{
This function generates a plot that summarizes the antibodies used in a CycifStack object.
It visualizes the antibodies across different samples and cycles.
}
\seealso{
\code{\link{abs_list}}
}
