% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{AbsSummary}
\alias{AbsSummary}
\title{Split a string}
\usage{
AbsSummary(x, show.cycles.in.row = FALSE, ...)
}
\arguments{
\item{x}{An object of Cycif class.}

\item{show.cycles.in.row}{logical. FALSE by default. If TRUE, the output plot shows the number
of cycles each sample was stained for.}

\item{...}{Arguments passed to image() function.}
}
\description{
Split a string
}
