% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{barplotPosCells}
\alias{barplotPosCells}
\alias{barplotPosCells,Cycif-method}
\title{Get stats for positive cells after gating each abs}
\usage{
barplotPosCells(x, ...)

\S4method{barplotPosCells}{Cycif}(x, type = c("one", "two"), mar, ...)
}
\description{
Get stats for positive cells after gating each abs
}
