% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotPosCells.R
\name{barplotPosCells}
\alias{barplotPosCells}
\title{Get stats for positive cells after gating each abs}
\usage{
barplotPosCells(x, ...)
}
\description{
Get stats for positive cells after gating each abs
}
