% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{CellNeighborhood-class}
\alias{CellNeighborhood-class}
\title{Class "CellNeighborhood" - Cell Neighborhood Object}
\description{
An object of class "CellNeighborhood" represents the results of the computeRCN function
for a Cycif or CycifStack object, providing information about the cell neighborhood analysis.
}
\section{Slots}{

\describe{
\item{\code{within.rois}}{A logical vector indicating whether each cell is within a region of interest (ROI).}

\item{\code{focused.cts}}{A character vector of the focused cell types around which Cell Neighborhood is considered in the analysis.}

\item{\code{cts.in.rcn}}{A character vector of cell types considered in the neighborhood analysis.}

\item{\code{n.cells.selected}}{An integer indicating the number of cells selected for the analysis.}

\item{\code{frnn}}{A frNN object containing information about cell neighbors.}

\item{\code{exp}}{A data.table containing expression data for selected cells.}

\item{\code{is.selected}}{A logical vector indicating whether each cell is selected.}

\item{\code{rcn.count}}{A matrix containing the counts of cell types in the neighborhood.}

\item{\code{rcn.freq}}{A matrix containing the frequency of cell types in the neighborhood.}
}}

\seealso{
\code{\link{computeRCN}}
}
