% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{CellNeighborhood-class}
\alias{CellNeighborhood-class}
\title{Class "CellNeighborhood" - Cell Neighborhood Object}
\description{
An object of class "CellNeighborhood" represents the results of the computeCN function
for a Cycif or CycifStack object, providing information about the cell neighborhood analysis.
}
\section{Slots}{

\describe{
\item{\code{within.rois}}{A logical vector indicating whether each cell is within a region of interest (ROI).}

\item{\code{cts.in.rcn}}{A character vector of cell types considered in the neighborhood analysis.}

\item{\code{n.cells.selected}}{An integer indicating the number of cells selected for the analysis.}

\item{\code{smpls}}{A character vector containing sample names.}

\item{\code{frnn}}{A frNN object containing information about cell neighbors.}

\item{\code{n.neighbors}}{An integer indicating the number of neighbors (including self).}

\item{\code{exp.per.ct.cn}}{A data.table containing expression data for each cell type in the neighborhood.}

\item{\code{exp.per.cn}}{A data.table containing expression data for each cell neighborhood.}

\item{\code{is.selected}}{A logical vector indicating whether each cell is selected.}

\item{\code{rcn.count}}{A data.frame containing the counts of cell types in the neighborhood.}

\item{\code{rcn.freq}}{A data.frame containing the frequency of cell types in the neighborhood.}

\item{\code{dist2tumorBorder}}{A numeric vector containing the distance of each cell to tumor border.}

\item{\code{mclustda}}{A list containing the results of the mclustDA function.}
}}

\seealso{
\code{\link{computeCN}}
}
