% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{cellTypeFrequency}
\alias{cellTypeFrequency}
\title{show cell type frequency table}
\usage{
cellTypeFrequency(x, ...)
}
\description{
Perform a cell type calling function and set cell types in a Cycif or CycifStack object
}
