% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycif.R
\name{cumUsedCells}
\alias{cumUsedCells}
\title{Show available cells at each cycle}
\usage{
cumUsedCells(x,within.rois=TRUE)
}
\arguments{
\item{x}{A Cycif object}

\item{cumulative}{logical. If TRUE, available cells through cycles 1 to n are computed.
If FALSE, available cells at each cycle are computed.}

\item{ratio}{Logical. If TRUE, show ratio of available and unavailable cells
at each cycle. If FALSE, actual actual number is provided.}
}
\value{
cumUsedCells() returns a matrix where rows correspond to cells
and columns correspond to cycles. statUsedCells() returns a table sumarizing
the numbers of available and unavailable cells at each cycle.
}
\description{
Show available cells at each cycle
}
