% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cycif-utils.R, R/nCycles.R
\name{names,Cycif-method}
\alias{names,Cycif-method}
\alias{abs_list}
\alias{abs_params}
\alias{abs_params<-}
\alias{dna}
\alias{nCells}
\alias{xys}
\alias{segProp}
\alias{dna_thres}
\alias{used_cells}
\alias{within_rois}
\alias{within_rois<-}
\alias{cell_type}
\alias{cell_type<-}
\alias{nCycles}
\alias{nCycles<-}
\title{Set nCycles - note this should run before cell type calling}
\usage{
\S4method{names}{Cycif}(x)

abs_list(x)

abs_params(x)

abs_params(x, ...) <- value

dna(x)

nCells(x)

xys(x)

segProp(x)

dna_thres(x)

used_cells(x)

within_rois(x)

within_rois(x, ...) <- value

cell_type(x)

cell_type(x, ...) <- value

nCycles(x)

nCycles(x, ...) <- value
}
\description{
Set nCycles - note this should run before cell type calling
}
