% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycif.R
\name{Cycif}
\alias{Cycif}
\alias{show,Cycif-method}
\title{Instantiate and show a Cycif object}
\usage{
Cycif(filename, file.suffix = "", col.suffix = "_cellMask")

\S4method{show}{Cycif}(object)
}
\arguments{
\item{filename}{quantification file (.csv)}

\item{file.suffix}{a common suffix in each file name. If applicable, the suffix is
removed from the channel names.}

\item{col.suffix}{a common suffix of each channel. If applicable, the suffix is
removed from the channel names.}

\item{object}{a Cycif object}
}
\description{
Instantiate and show a Cycif object
}
