% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycif.R
\name{Cycif}
\alias{Cycif}
\alias{show,Cycif-method}
\title{Instantiate and show a Cycif object}
\usage{
Cycif(
  ft_filename,
  path = "",
  mask_type = c("cellRing", "cell"),
  mcmicro = FALSE,
  use_scimap = FALSE
)

\S4method{show}{Cycif}(object)
}
\arguments{
\item{ft_filename}{quantification file (.csv)}

\item{path}{path to the dir containing ft_filename}

\item{mask_type}{a common mask_type of each file name or channel. If applicable, the mask_type is
removed from the channel names.}

\item{object}{a Cycif object}
}
\description{
Instantiate and show a Cycif object
}
