% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CycifStack-utils.R, R/nCycles.R
\name{nSamples}
\alias{nSamples}
\alias{names,CycifStack-method}
\alias{abs_list,CycifStack-method}
\alias{maxCycles}
\alias{nCells,CycifStack-method}
\alias{length,CycifStack-method}
\alias{nCycles,CycifStack-method}
\title{Accessing slots in a CellTypeCycifStack object}
\usage{
nSamples(x)

\S4method{names}{CycifStack}(x)

\S4method{abs_list}{CycifStack}(x)

maxCycles(x)

\S4method{nCells}{CycifStack}(x)

\S4method{length}{CycifStack}(x)

\S4method{nCycles}{CycifStack}(x)
}
\arguments{
\item{x}{A CycifStack object}
}
\description{
Accessing slots in a CellTypeCycifStack object
}
