% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycifStack.R
\name{CycifStack}
\alias{CycifStack}
\alias{show,CycifStack-method}
\title{Instantiate and show a CycifStack object}
\usage{
CycifStack(filenames, path, suffix = "_cellMask")

\S4method{show}{CycifStack}(object)
}
\arguments{
\item{filenames}{quantification file (.csv)}

\item{suffix}{a common suffix of each channel. If applicable, the suffix is
removed from the channel names.}

\item{object}{a CycifStack object}
}
\description{
Instantiate and show a CycifStack object
}
