% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycifStack.R
\name{[,CycifStack-method}
\alias{[,CycifStack-method}
\alias{[[,CycifStack-method}
\alias{[<-,CycifStack-method}
\alias{[[<-,CycifStack-method}
\title{Subsetting samples from a CycifStack object}
\usage{
\S4method{[}{CycifStack}(x, i = NULL)

\S4method{[[}{CycifStack}(x, i = NULL)

\S4method{[}{CycifStack}(x, i) <- value

\S4method{[[}{CycifStack}(x, i) <- value
}
\arguments{
\item{x}{A CycifStack object}

\item{i}{A numeric or character vector. If numeric, the values are integers
representing indices for the Cycif objects. If character, names of the Cycif
objects.}

\item{value}{A Cycif object to be inserted or replaced.}
}
\description{
Subsetting samples from a CycifStack object
}
