% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{defineCellTypes}
\alias{defineCellTypes}
\alias{defineCellTypes,data.frame-method}
\alias{defineCellTypes,Cycif-method}
\alias{defineCellTypes,CycifStack-method}
\title{Define Cell Types and set them in a Cycif or CycifStack object}
\usage{
defineCellTypes(x, ...)

\S4method{defineCellTypes}{data.frame}(
  x,
  ctype,
  cstate,
  p_thres = 0.5,
  mc.cores = 4,
  prioritized.celltypes = NULL,
  ...
)

\S4method{defineCellTypes}{Cycif}(
  x,
  ctype,
  cstate,
  ct_name = "default",
  p_thres = 0.5,
  mc.cores = 4,
  overwrite = FALSE,
  prioritized.celltypes = NULL,
  ...
)

\S4method{defineCellTypes}{CycifStack}(
  x,
  ctype,
  cstate,
  ct_name = "default",
  p_thres = 0.5,
  mc.cores = 4,
  overwrite = FALSE,
  prioritized.celltypes = NULL,
  ...
)
}
\arguments{
\item{x}{A Cycif or CycifStack object.}

\item{...}{Additional arguments (currently unused).}

\item{ctype}{A data.frame containing cell type definition.}

\item{cstate}{A data.frame containing cell state definition.}

\item{p_thres}{Numerical value between 0 and 1. A probability that corresponds to a threshold intensity. Default is 0.5.}

\item{mc.cores}{Number of CPU cores to use for parallel processing. Default is 4.}

\item{ct_name}{Name of the cell types.}

\item{overwrite}{Logical, if TRUE, overwrite existing cell type definitions with the same name. Default is FALSE.}
}
\value{
A modified Cycif or CycifStack object with updated cell type information.
}
\description{
This function performs a cell type calling operation and sets cell types in a Cycif or CycifStack object.
}
\details{
The \code{defineCellTypes} function performs a cell type calling operation and sets cell types in a Cycif or CycifStack object. It takes as input a \code{ctype} data.frame, a \code{cstate} data.frame, and optional parameters for thresholding and parallel processing. Cell types can be defined at different hierarchy levels, and the results can be stored under a specified \code{ct_name}.

If \code{ct_name} already exists and \code{overwrite} is set to FALSE, the function will not overwrite the existing cell type definitions and will issue a warning.
}
\seealso{
\code{\link{CellTypeSkeleton}}, \code{\link{cyApply}}
}
