% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineCellTypes.R
\name{defineCellTypes}
\alias{defineCellTypes}
\title{Perform a cell type calling function and set cell types in a Cycif or CycifStack object}
\usage{
defineCellTypes(x, ...)
}
\description{
Perform a cell type calling function and set cell types in a Cycif or CycifStack object
}
