% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis.R
\name{defineTumorBed}
\alias{defineTumorBed}
\alias{defineTumorBed,Cycif-method}
\title{Identify Tumor bed based on dbscan clusters}
\usage{
defineTumorBed(x, ...)

\S4method{defineTumorBed}{Cycif}(
  x,
  strict = FALSE,
  dth,
  parallel = TRUE,
  min.pts = 2,
  n.cores = 7
)
}
\arguments{
\item{x}{A Cycif obj}

\item{strict}{logical. if cell_types should be strict.}

\item{dth}{radius for dbscan (eps)}

\item{parallel}{logical. The default is TRUE.}

\item{min.pts}{minimum points for dbscan}

\item{n.cores}{number of cores.}
}
\description{
Identify Tumor bed based on dbscan clusters
}
