% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimPlot.R
\name{dimPlot}
\alias{dimPlot}
\alias{dimPlot,Cycif-method}
\alias{dimPlot,CycifStack-method}
\title{Plot 2D-coordinates computed by UMAP or T-SNE}
\usage{
dimPlot(x, ...)

\S4method{dimPlot}{Cycif}(
  x,
  pch = ".",
  type = c("cell_type", "exp"),
  celltype = cts,
  ab,
  main,
  leg = TRUE,
  p_thres = 0.4,
  ...
)

\S4method{dimPlot}{CycifStack}(
  x,
  pch = ".",
  type = c("cell_type", "smpl", "exp"),
  celltype,
  ab,
  main,
  leg = TRUE,
  p_thres = 0.4,
  ...
)
}
\arguments{
\item{x}{A CycifStack object}

\item{...}{other arguments  passed to plot()}

\item{pch}{pch passed to plot()}

\item{type}{character, indicating how cells are colored. 'cell_type' is cell types
computed prior to the plotting, 'smpl' show the samples the cells derived from,
and 'exp' is expression of one antibody. When 'cell_type' and 'exp', additional
parameters needs to be provided.}

\item{celltype}{a factor with the length of cell number, indicating cell types computed
elsewhere. This needs to be provided when type is 'cell_type'.}

\item{ab}{character, indicating one channel/antibody. This needs to be provided when
type is 'exp'.}

\item{main}{main passed to plot()}

\item{uniq.cols}{color sets used for the plot()}
}
\description{
Plot 2D-coordinates computed by UMAP or T-SNE
}
