% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis.R
\name{dist2tumorBorder}
\alias{dist2tumorBorder}
\alias{dist2tumorBorder,Cycif-method}
\title{Calculate the Distance to Tumor Border in CyCIF data.}
\usage{
dist2tumorBorder(x, ...)

\S4method{dist2tumorBorder}{Cycif}(
  x,
  n.cores = 7,
  dth,
  th,
  minPts = 3,
  concavity = 0.8,
  plot = FALSE,
  ...
)
}
\arguments{
\item{x}{A CyCIF object.}

\item{...}{Additional arguments to be passed.}

\item{n.cores}{Integer, number of CPU cores to use for parallel processing.}

\item{dth}{Numeric, the distance threshold used for tumor border detection.}

\item{minPts}{Integer, the minimum number of points required to form a cluster in DBSCAN.}

\item{concavity}{a relative measure of concavity. 1 results in a relatively detailed shape, Infinity results in a convex hull. You can use values lower than 1, but they can produce pretty crazy shapes (\code{concaveman}).}
}
\value{
A list containing the calculated distances and tumor border polygons.
}
\description{
This function calculates the distance of each cell in CyCIF data to the nearest tumor border. The tumor border is defined based on the spatial arrangement of tumor cells within the dataset.
}
\details{
The \code{dist2tumorBorder} function calculates the distance of each cell in the CyCIF dataset to the nearest tumor border. It performs the following steps:
\enumerate{
\item It extracts the coordinates of cells from the CyCIF data, considering the spatial arrangement of tumor cells.
\item It identifies tumor cells and their clusters using DBSCAN clustering based on the specified distance threshold (\code{dth}).
\item It computes the tumor border polygons by using the Concaveman algorithm on each tumor cell cluster.
\item It determines if each non-tumor cell is inside or outside of the tumor border polygons using the \code{point.in.polygon} function.
\item It calculates the distance of each cell to the nearest tumor border, considering whether the cell is inside or outside the tumor border.
\item The function returns a list containing the calculated distances and the tumor border polygons.
}
}
\seealso{
\code{\link{concaveman}}, \code{\link{point.in.polygon}}, \code{\link{dbscan}} \code{\link{gDistance}}
}
