% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis.R
\name{distAdjacentCells}
\alias{distAdjacentCells}
\alias{distAdjacentCells,Cycif-method}
\alias{distAdjacentCells,CycifStack-method}
\title{Calculate the Distance Between Adjacent Cells in CyCIF Data.}
\usage{
distAdjacentCells(x, ...)

\S4method{distAdjacentCells}{Cycif}(x)

\S4method{distAdjacentCells}{CycifStack}(x)
}
\description{
Calculate the Distance Between Adjacent Cells in CyCIF Data.
}
