% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dna}
\alias{dna}
\title{Get the DNA channels in a Cycif object}
\usage{
dna(x)
}
\arguments{
\item{x}{A Cycif object.}
}
\value{
A character vector containing the names of DNA channels in the Cycif object.
}
\description{
The \code{dna} function returns the DNA channels in a Cycif object.
}
\examples{
\dontrun{
dna(cycif_object)
}

}
\seealso{
\code{\link{nSamples}}: Get the number of samples in a CycifStack object.
\code{\link{names}}: Get the names of samples in a CycifStack object.
\code{\link{maxCycles}}: Get the maximum number of cycles in a CycifStack object.
\code{\link{within_rois}}: Check if samples in a CycifStack are within ROIs.
\code{\link{length}}: Equivalent to \code{nSamples} for CycifStack objects.
}
