% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnaFilter.R
\name{dnaFilter}
\alias{dnaFilter}
\title{Filtering out unavaialble cells based on nuclei staining}
\usage{
dnaFilter(x, ...)
}
\arguments{
\item{x}{A Cycif object}

\item{manual}{logical. If TRUE, lower and upper limits of DNA intensity should be chosen
interactively by clicking the limit values in the heatmap. FALSE by default.}

\item{ratio}{logical. If TRUE, the ratio of DNA intensity between each cycle and cycle 0 will be used.
If FALSE, raw values of DNA intensity will be used.}

\item{n}{numeric. The number of breaks in the histogram.}

\item{n1}{numeric. indices from 1 through n1 will be plugged into loess().}

\item{show.only}{logical. If TRUE, only show the summary of the plots but don't attempt to filter cells.}
}
\description{
Filtering out unavaialble cells based on nuclei staining
}
