% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{expandLineageDef}
\alias{expandLineageDef}
\title{internally used only within CellTypeDefault, so don't export}
\usage{
expandLineageDef(ctype, cstate, ctype.full = TRUE)
}
\description{
internally used only within CellTypeDefault, so don't export
}
