% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{exprs}
\alias{exprs}
\title{Show a raw or normalized protein expression matrix}
\usage{
exprs(x, ...)
}
\arguments{
\item{x}{A Cycif or CycifStack object.}

\item{type}{type. If "raw", a raw expression matrix is shown. If "normalized",
a normalized expression matrix will be shown. The method of normalization is
specified by "log" or "LogTh" in the normalize function provided previously.}

\item{silent}{logical. If FALSE, the method of normalization was shown in
the error prompt.}
}
\description{
Show a raw or normalized protein expression matrix
}
