% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{file_paths-class}
\alias{file_paths-class}
\title{File Paths Class}
\description{
This class represents file paths associated with CyCIF data processed with MCMICRO for a single sample.
}
\details{
The \code{file_paths} class is used to store information about the file paths for different steps of CyCIF data processing for a single sample.
}
\section{Slots}{

\describe{
\item{\code{sample_name}}{A character vector specifying the sample name.}
\item{\code{root_dir}}{A character vector indicating the root directory where data files are stored.}
\item{\code{mask_type}}{A character vector representing the mask type used for processing only 'cell' and 'cellRing' are accepted.}
\item{\code{registration}}{A character vector indicating the file path for registration data (ome.tif file).}
\item{\code{segmentation}}{A character vector indicating the file path for segmentation data (ome.tif file).}
\item{\code{quantification}}{A character vector indicating the file path for quantification data, also known as spatial feature table (csv file).}
}
}

\examples{
# Creating a file_paths object
file_paths_object <- new("file_paths",
                         sample_name = "Sample1",
                         root_dir = "/path/to/data/",
                         mask_type = "cell",
                         registration = "/path/to/registration/file",
                         segmentation = "/path/to/segmentation/file",
                         quantification = "/path/to/quantification/file")

# Accessing slots
sample_name <- file_paths_object@sample_name
root_dir <- file_paths_object@root_dir
mask_type <- file_paths_object@mask_type
registration <- file_paths_object@registration
segmentation <- file_paths_object@segmentation
quantification <- file_paths_object@quantification

}
\seealso{
Other classes: \code{\link{Cycif}}, \code{\link{CycifStack}}, \code{\link{CellTypes}}
}
