% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycif.R, R/cycifstack.R, R/celltype.R
\name{gates}
\alias{gates}
\alias{gates,Cycif-method}
\alias{gates,CycifStack-method}
\alias{gates,CellTypeCycif-method}
\title{Show gates in a Cycif/CycifStack/CellTypeCycif/CellTypeCicyfStack object}
\usage{
gates(x)

\S4method{gates}{Cycif}(x)

\S4method{gates}{CycifStack}(x)

\S4method{gates}{CellTypeCycif}(x)
}
\description{
Show gates in a Cycif/CycifStack/CellTypeCycif/CellTypeCicyfStack object
}
