% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{getGates}
\alias{getGates}
\alias{getGates,Cycif-method}
\alias{getGates,CycifStack-method}
\alias{setGates}
\alias{setGates,Cycif-method}
\alias{setGates,CycifStack-method}
\title{Get or set gates information in a Cycif or CycifStack object}
\usage{
getGates(x)

\S4method{getGates}{Cycif}(x)

\S4method{getGates}{CycifStack}(x)

setGates(x, ...)

\S4method{setGates}{Cycif}(x, gates.df, run_normalize = TRUE, p_thres = 0.5, trim = 0.001)

\S4method{setGates}{CycifStack}(x, gates.df, run_normalize = TRUE, p_thres = 0.5, trim = 0.001)
}
\arguments{
\item{x}{A Cycif or CycifStack object.}

\item{...}{Additional arguments specific to the get or set operation.}
}
\value{
If used as \code{getGates(x)}, it returns the gates information.
If used as \code{setGates(x, ...)}, it sets the gates information and returns the modified object.
}
\description{
This function allows you to get or set gates information in a Cycif or CycifStack object.
}
\examples{
# Get gates information from a Cycif object
ct_obj <- new("Cycif", ...)
getGates(ct_obj)

# Set gates information in a Cycif object
setGates(ct_obj, gates_data)

# Get gates information from a CycifStack object
ct_stack_obj <- new("CycifStack", ...)
getGates(ct_stack_obj)

# Set gates information in a CycifStack object
setGates(ct_stack_obj, gates_data)

}
