% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{h3}
\alias{h3}
\title{3D Heatmap Plot}
\usage{
h3(
  x,
  Rowv = NULL,
  Colv = if (symm) "Rowv" else NULL,
  distfun = dist,
  distfunC,
  distfunR,
  balanceColor = F,
  ColSideLabs,
  RowSideLabs,
  showColDendro = T,
  showRowDendro = T,
  col = colorRampPalette(c("navy", "white", "firebrick3"))(1024),
  legendfun,
  method = "complete",
  ColAxisColors = 0,
  RowAxisColors = 0,
  hclustfun = hclust,
  reorderfun = function(d, w) reorder(d, w),
  add.expr,
  symm = FALSE,
  revC = identical(Colv, "Rowv"),
  scale = c("row", "column", "none"),
  na.rm = TRUE,
  ColSideFun,
  ColSideAnn,
  ColSideWidth = 0.4,
  ColSideCut,
  colorCell,
  highlightCell,
  file = "heatmap3.pdf",
  topN = NA,
  filterFun = sd,
  returnDistMatrix = FALSE,
  margins = c(5, 5),
  ColSideColors,
  RowSideColors,
  cexRow = 0.2 + 1/log10(nrow(x)),
  cexCol = 0.2 + 1/log10(ncol(x)),
  lasRow = 2,
  lasCol = 2,
  labRow = NULL,
  labCol = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  na.color = "grey",
  keep.dendro = FALSE,
  verbose = getOption("verbose"),
  useRaster = if (ncol(x) * nrow(x) >= 50000) TRUE else FALSE,
  ...
)
}
\arguments{
\item{x}{A numeric matrix containing the data to be plotted.}

\item{Rowv}{Specifies the row dendrogram or clustering (default is NULL).}

\item{Colv}{Specifies the column dendrogram or clustering (default is NULL).}

\item{distfun}{A function to calculate the distance matrix for rows (default is dist).}

\item{distfunC}{A function to calculate the distance matrix for columns (default is NULL).}

\item{distfunR}{A function to calculate the distance matrix for rows (default is NULL).}

\item{balanceColor}{Logical, indicating whether to balance colors (default is FALSE).}

\item{ColSideLabs}{Labels for the columns' side (default is NULL).}

\item{RowSideLabs}{Labels for the rows' side (default is NULL).}

\item{showColDendro}{Logical, indicating whether to show the column dendrogram (default is TRUE).}

\item{showRowDendro}{Logical, indicating whether to show the row dendrogram (default is TRUE).}

\item{col}{A color palette for the heatmap (default is a gradient from navy to firebrick3).}

\item{legendfun}{A custom function to create the legend (default is NULL).}
}
\description{
This function creates a 3D heatmap plot to visualize a matrix of data.
}
\details{
\itemize{
\item The \code{h3} function creates a 3D heatmap plot to visualize a matrix of data.
\item It allows customization of various aspects of the plot, including colors, labels, and dendrograms.
\item Additional functionalities like distance matrix calculations and custom legends are available.
}
}
