% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{ld_coords}
\alias{ld_coords}
\alias{ld_coords,Cycif-method}
\alias{ld_coords,CycifStack-method}
\title{Get UMAP Coordinates or Clustering Results}
\usage{
ld_coords(x, ...)

\S4method{ld_coords}{Cycif}(x, ld_name)

\S4method{ld_coords}{CycifStack}(x, ld_name)
}
\arguments{
\item{x}{A 'Cycif' or 'CycifStack' object containing UMAP embeddings or clustering results.}

\item{ld_name}{The name of the UMAP embedding or clustering result to retrieve.}
}
\value{
UMAP coordinates or clustering results specified by 'ld_name'.
}
\description{
This function retrieves UMAP coordinates or clustering results associated with a specified name from a 'Cycif' or 'CycifStack' object.
}
\details{
The 'ld_coords' function allows you to obtain UMAP coordinates or clustering outcomes stored within a 'Cycif' or 'CycifStack' object. You need to specify the name of the UMAP embedding or clustering result you want to retrieve using the 'ld_name' parameter. If the specified 'ld_name' does not exist in the object, an error will be raised.
}
