% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycif.R
\name{list2Cycif}
\alias{list2Cycif}
\title{Converting a list to a Cycif object}
\usage{
list2Cycif(x)
}
\arguments{
\item{x}{A list object}
}
\value{
A Cycif object
}
\description{
Converting a list to a Cycif object
}
