% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycifstack.R
\name{list2CycifStack}
\alias{list2CycifStack}
\title{Converting a list to a CycifStack object}
\usage{
list2CycifStack(x)
}
\arguments{
\item{x}{A list or a Cycif object (if Cycif, the length should be one)}
}
\value{
A CycifStack object
}
\description{
Converting a list to a CycifStack object
}
