% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{lowDimPlot}
\alias{lowDimPlot}
\alias{lowDimPlot,Cycif-method}
\alias{lowDimPlot,CycifStack-method}
\title{Plot 2D-coordinates computed by UMAP or T-SNE}
\usage{
lowDimPlot(x, ...)

\S4method{lowDimPlot}{Cycif}(
  x,
  ld_name,
  plot_type = c("celltype", "cluster", "exp"),
  ab,
  na.col = "grey80",
  uniq.cols,
  with.labels = TRUE,
  leg = TRUE,
  pch = ".",
  main,
  p_thres = 0.5,
  mar,
  cex.labs = 1,
  cex.leg = 1,
  cex.main = 2,
  ...
)

\S4method{lowDimPlot}{CycifStack}(
  x,
  ld_name,
  plot_type = c("celltype", "cluster", "exp"),
  ab,
  na.col = "grey80",
  uniq.cols,
  with.labels = TRUE,
  leg = TRUE,
  cts.names,
  pal.name,
  cex.main = 2,
  pch = ".",
  main,
  p_thres = 0.5,
  mar,
  cex.labs = 1,
  cex.leg = 0.5,
  cex = cex,
  ...
)
}
\arguments{
\item{x}{A CycifStack object}

\item{...}{other arguments  passed to plot()}

\item{ab}{character, indicating one channel/antibody. This needs to be provided when
type is 'exp'.}

\item{uniq.cols}{color sets used for the plot()}

\item{pch}{pch passed to plot()}

\item{main}{main passed to plot()}

\item{type}{character, indicating how cells are colored. 'cell_type' is cell types
computed prior to the plotting, 'smpl' show the samples the cells derived from,
and 'exp' is expression of one antibody. When 'cell_type' and 'exp', additional
parameters needs to be provided.}

\item{celltype}{a factor with the length of cell number, indicating cell types computed
elsewhere. This needs to be provided when type is 'cell_type'.}
}
\description{
Plot 2D-coordinates computed by UMAP or T-SNE
}
