% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{lowDimPlot}
\alias{lowDimPlot}
\alias{lowDimPlot,Cycif-method}
\alias{lowDimPlot,CycifStack-method}
\title{Create Low-Dimensional Plot}
\usage{
lowDimPlot(x, ...)

\S4method{lowDimPlot}{Cycif}(
  x,
  ld_name,
  plot_type = c("cell_type", "cluster", "exp", "smpl"),
  ab,
  uniq.cols,
  with.labels = TRUE,
  ct_name = "default",
  used.smpls,
  xlab = "",
  ylab = "",
  used.cts,
  cex.main = 2,
  main,
  p_thres = 0.5,
  mar,
  cex.labs = 1,
  cex.leg = 0.5,
  cex = 0.3,
  ...
)

\S4method{lowDimPlot}{CycifStack}(
  x,
  ld_name,
  plot_type = c("cell_type", "cluster", "exp", "smpl"),
  ab,
  uniq.cols,
  with.labels = TRUE,
  ct_name = "default",
  used.smpls,
  xlab = "",
  ylab = "",
  used.cts,
  cex.main = 2,
  main,
  p_thres = 0.5,
  mar,
  cex.labs = 1,
  cex.leg = 0.5,
  cex = 0.3,
  ...
)
}
\arguments{
\item{x}{A \code{Cycif} or \code{CycifStack} object containing protein expression data.}

\item{ld_name}{The name of the low-dimensional representation (e.g., UMAP or clustering).}

\item{plot_type}{The type of low-dimensional plot to create, one of "cell_type", "cluster", "exp", or "smpl".}

\item{ab}{The name of the protein to be used for coloring points (required when \code{plot_type} is "exp").}

\item{uniq.cols}{Vector of unique colors to use for plotting points (optional).}

\item{with.labels}{Logical, indicating whether to label points (default is TRUE).}

\item{ct_name}{The name of the cell type column (default is "default").}

\item{used.smpls}{Vector of samples to include in the plot.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{used.cts}{Vector of cell types to include in the plot.}

\item{cex.main}{Size of the main title text (default is 2).}

\item{main}{Title for the plot (default is auto-generated based on parameters).}

\item{p_thres}{Threshold for plotting (default is 0.5).}

\item{mar}{Margins of the plot.}

\item{cex.labs}{Size of label text (default is 1).}

\item{cex.leg}{Size of the legend text (default is 0.5).}

\item{cex}{Size of data point labels (default is 0.3).}
}
\description{
This function generates low-dimensional plots for protein expression data from a \code{Cycif} or \code{CycifStack} object.
}
\details{
\itemize{
\item The \code{lowDimPlot} function creates low-dimensional plots for protein expression data based on the specified \code{Cycif} or \code{CycifStack} object.
\item Users can choose from different types of low-dimensional plots, including those based on cell types, clusters, protein expression, or samples.
\item Various customization options are available for labeling, coloring, and scaling the plot.
}
}
