% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{maxCycles}
\alias{maxCycles}
\title{Get the maximum number of cycles used in a CycifStack object}
\usage{
maxCycles(x)
}
\arguments{
\item{x}{A CycifStack object.}
}
\value{
The maximum number of cycles in the CycifStack object.
}
\description{
The \code{maxCycles} function returns the maximum number of cycles in a CycifStack object.
}
\seealso{
\code{\link{nSamples}}: Get the number of samples in a CycifStack object.
\code{\link{names}}: Get the names of samples in a CycifStack object.
\code{\link{within_rois}}: Check if samples in a CycifStack are within ROIs.
}
