% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis.R
\name{meanExpRCN}
\alias{meanExpRCN}
\alias{meanExpRCN,data.frame-method}
\title{Compute Mean Expression Profiles per RCN Cluster}
\usage{
meanExpRCN(x, ...)

\S4method{meanExpRCN}{data.frame}(
  x,
  frnn,
  focused.cts = "Tumor",
  cts.in.rcn = levels(cell_types(x)$cell_types)[1:10],
  per.ct = TRUE,
  extrapolate = TRUE
)
}
\arguments{
\item{x}{A data frame containing expression data, typically from a CyCIF or similar dataset.}

\item{frnn}{An object containing RCN information, typically obtained from 'computeRCN'.}

\item{focused.cts}{A character vector specifying the cell types to focus on (e.g., "Tumor").}

\item{cts.in.rcn}{A character vector specifying the cell types to include in the RCN analysis.}

\item{per.ct}{A logical value indicating whether to compute mean expression profiles per RCN cluster (TRUE) or for the entire dataset (FALSE).}

\item{extrapolate}{A logical value indicating whether to extrapolate clustering results to the entire dataset (TRUE) or not (FALSE).}
}
\value{
A list of data frames containing mean expression profiles for specified cell types or features within RCN clusters.
}
\description{
This function computes the mean expression profiles for specified cell types
or features within Recurrent Cell Neighborhood (RCN) clusters. It takes a data frame,
typically containing expression data, and computes the mean expression values
for each RCN cluster based on the provided RCN information from 'computeRCN'.
The function allows you to focus on specific cell types and features and can
extrapolate the clustering results to the entire dataset if needed.
}
\details{
The 'meanExpRCN' function calculates the mean expression profiles for the specified cell types or features within RCN clusters. The function works as follows:
\itemize{
\item It takes a data frame 'x' containing expression data and an object 'frnn' containing RCN information obtained from the 'computeRCN' function.
\item You can specify the 'focused.cts' argument to select specific cell types to focus on during the analysis.
\item The 'cts.in.rcn' argument allows you to specify the cell types to include in the RCN analysis.
\item If 'per.ct' is set to TRUE, the function computes mean expression profiles per RCN cluster; otherwise, it computes mean expression profiles for the entire dataset.
\item The 'extrapolate' argument determines whether to extrapolate clustering results to the entire dataset based on RCN information.
The function returns a list of data frames containing mean expression profiles for the specified cell types or features within RCN clusters.
}
}
\seealso{
\code{\link{computeRCN}}, \code{\link{rcnClust}}
}
