% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycif.R, R/cycifstack.R
\name{nCycles}
\alias{nCycles}
\alias{nCycles,Cycif-method}
\alias{nCycles,CycifStack-method}
\title{#' #' Replace the values of within_rois for a Cycif object
#' #' @param object A cycif object.
#' #' @param value logical. If TRUE, the area within ROIs is concerned for the downstream
#' #' @export
#' within_rois <- function(object, value) {
#'   if(!is(object,"Cycif")){
#'     stop("this function applies only to a Cycif object")
#'   }
#'   object@within_rois <- value
#'   return(object)
#' }}
\usage{
nCycles(x)

\S4method{nCycles}{Cycif}(x)

\S4method{nCycles}{CycifStack}(x)
}
\description{
#' Replace the values of within_rois for a Cycif object
#'
#' @param obj A \code{\link{Cycif}} object.
#' @param value A named list of data.frames, with names corresponding to
#'   the names of \code{object}'s channels.
#'
#' @return An updated \code{\link{Cycif}} object.
#'
#' @usage
#' within_rois(x) <- value
#' @export
setReplaceMethod("within_rois",
signature(obj = "Cycif", value = "logical"),
function(obj, value) {
obj@name <- value
return(obj)
})
}
