% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{normalize}
\alias{normalize}
\alias{normalize,Cycif-method}
\alias{transform}
\title{Show a raw or normalized protein expression matrix}
\usage{
normalize(x, ...)

\S4method{normalize}{Cycif}(x, method = c("log", "logTh", "invlog"), trim = 0.001, p_thres = 0.5)

transform(
  r,
  method = c("log", "logTh", "Th", "invlog"),
  th,
  p_thres = 0.5,
  trim = 0.001
)
}
\arguments{
\item{x}{A Cycif or CycifStack object.}

\item{method}{Either "log" or "logTh". "log" indicates that
the raw values are transformed using log1p() function. "logTh" method
further trim the outliers outside outside [trim,1-trim] quantiles,
where trm is specified by trim parameter. Outlier-trimmed, log1p-transformed
raw values are further transformed such that the value indicated by the threshold,
th (in the case of transform() function) or threshold(x) (in the case of
normalize() function) will be set to p_thres. Values below or above the threshold
are linearly transformed to values between 0 and p_thres, or values between p_thres and 1,
respectively.}

\item{r}{A numeric scholar indicating a raw expression value.}

\item{th}{A numeric scholar. A user-provided threshold for a raw expression value
for each protein for each sample.}
}
\description{
Show a raw or normalized protein expression matrix
}
