% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nSamples}
\alias{nSamples}
\alias{nSamples,CycifStack-method}
\alias{length,CycifStack-method}
\title{Get the number of samples in a CycifStack object}
\usage{
nSamples(x)

\S4method{nSamples}{CycifStack}(x)

\S4method{length}{CycifStack}(x)
}
\arguments{
\item{x}{A CycifStack object.}
}
\value{
The number of samples in the CycifStack object.
}
\description{
The \code{nSamples} function returns the number of samples in a CycifStack object.
}
\seealso{
\code{\link{names}}: Get the names of samples in a CycifStack object.
\code{\link{maxCycles}}: Get the maximum number of cycles in a CycifStack object.
\code{\link{within_rois}}: Check if samples in a CycifStack are within ROIs.
}
