% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycifStack.R
\name{nSamples}
\alias{nSamples}
\title{Accessing slots in a CellTypeCycifStack object}
\usage{
nSamples(x)
}
\arguments{
\item{x}{A CycifStack object}
}
\description{
Accessing slots in a CellTypeCycifStack object
}
