% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotAvailCellOnSlide}
\alias{plotAvailCellOnSlide}
\alias{plotAvailCellOnSlide,Cycif-method}
\title{Plot Available Cells on Slide}
\usage{
plotAvailCellOnSlide(x, ...)

\S4method{plotAvailCellOnSlide}{Cycif}(
  x,
  upside.down = TRUE,
  ncycle,
  mfrow = c(3, 3),
  mar = c(0, 0, 4, 0),
  legend = TRUE,
  main = names(x),
  cex.title = 1,
  uniq.cols = c(lost = "grey90", dropped = "blue", available = "black", bunched = "red"),
  legend.cex = 2,
  xlab = "",
  ylab = "",
  ...
)
}
\arguments{
\item{x}{A Cycif object.}

\item{...}{Additional graphical parameters to customize the individual plots.}

\item{upside.down}{Should the plot be upside-down? (default is TRUE)}

\item{ncycle}{The number of cycles to consider for plotting (default is determined by the data).}

\item{mfrow}{Number of rows and columns in the grid of plots (default is c(3, 3)).}

\item{mar}{Margins for each individual plot (default is c(0, 0, 4, 0)).}

\item{legend}{Should a legend be included in the plot? (default is TRUE)}

\item{main}{Main title for the plot (default is the names of the Cycif object).}

\item{cex.title}{Size of the main title text (default is 1).}

\item{uniq.cols}{Colors for different cell states (default colors).}

\item{legend.cex}{Size of the legend text (default is 2).}

\item{xlab}{Label for the x-axis (default is empty).}

\item{ylab}{Label for the y-axis (default is empty).}
}
\description{
This function creates a plot that visualizes the available cells on a slide in Cycif data.
}
\details{
\itemize{
\item The \code{plotAvailCellOnSlide} function creates a grid of plots, each representing the available cells on a slide for a specific cycle.
\item The plots show the spatial distribution of available cells in different colors, with an optional legend.
\item You can customize the appearance of the plots and the legend using various graphical parameters.
}
}
