% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotUsedCellRatio}
\alias{plotUsedCellRatio}
\alias{plotUsedCellRatio,Cycif-method}
\alias{plotUsedCellRatio,CycifStack-method}
\title{Plot the ratio of used cells}
\usage{
plotUsedCellRatio(x, ...)

\S4method{plotUsedCellRatio}{Cycif}(
  x,
  cumulative = TRUE,
  ncycle,
  mar = c(5, 5, 4, 10),
  main = "# cells attached on slide",
  ...
)

\S4method{plotUsedCellRatio}{CycifStack}(
  x,
  cumulative = TRUE,
  ncycle,
  mar = c(5, 5, 4, 10),
  main = "# cells attached on slide",
  smpl.cols,
  ncol = 1,
  leg.cex = 0.8,
  use_rois = TRUE,
  ...
)
}
\description{
This function creates a plot of the ratio of used cells across channels.
}
