% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis.R
\name{ratioTumorStroma}
\alias{ratioTumorStroma}
\alias{ratioTumorStroma,Cycif-method}
\title{Calculate the Tumor-to-Stroma Ratio in CyCIF data.}
\usage{
ratioTumorStroma(x, ...)

\S4method{ratioTumorStroma}{Cycif}(x, stroma.cts, ...)

\S4method{ratioTumorStroma}{Cycif}(x, stroma.cts, ...)
}
\arguments{
\item{x}{A CyCIF or CyCIFStack object.}

\item{...}{Additional arguments to be passed to the \code{\link{defineTumorBorder}} function.}

\item{stroma.cts}{Character vector, specifying the stroma cell types to consider for TSR calculation. If not provided, stroma cell types are automatically determined.}
}
\value{
A numeric vector or data frame containing the Tumor-to-Stroma Ratios (TSR) for each cell type in the dataset.
}
\description{
This function calculates the Tumor-to-Stroma Ratio (TSR) within CyCIF data for either a single CyCIF object or a CyCIFStack object. The TSR is calculated based on the presence of tumor cells within defined tumor bed regions and the specified stroma cell types.
}
\details{
The Tumor-to-Stroma Ratio (TSR) is calculated based on the presence of tumor cells within defined tumor bed regions and the specified stroma cell types. The function performs the following steps:
\enumerate{
\item If the input is a CyCIF object, it calculates the TSR for that object. If the input is a CyCIFStack object, it calculates the TSR for each CyCIF object in the stack.
\item It determines the tumor bed regions within the CyCIF data using the \code{\link{defineTumorBorder}} function.
\item It counts the number of cells of each cell type within the tumor bed regions.
\item If the \code{stroma.cts} parameter is not provided, stroma cell types are automatically determined based on unique cell types in the dataset, excluding "Tumor" and "outOfROI" cell types.
\item It calculates the overall TSR (r.all) as the ratio of stroma cells to all cells within the tumor bed.
\item It calculates the TSR for each specified stroma cell type, considering the ratio of that cell type to all cells within the tumor bed.
\item For CyCIFStack objects, the function returns a data frame containing TSR values for each CyCIF object in the stack.
}
}
\seealso{
\code{\link{defineTumorBorder}}
}
