% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis.R
\name{rcnClust}
\alias{rcnClust}
\alias{rcnClust,data.frame-method}
\title{Cluster and Sort Recurrent Cell Neighbors (RCN) for CyCIF or CyCIFStack Objects}
\usage{
rcnClust(frnn, ...)

\S4method{rcnClust}{data.frame}(
  frnn,
  g = 50,
  seed = 123,
  sort.by = "CD8T",
  sort.type = c("freq", "count"),
  sort.smpls = c("all", "selected"),
  data.type = c("ct_exp", "ct"),
  extrapolate = FALSE,
  mc.cores = 1
)
}
\arguments{
\item{frnn}{An object containing RCN information, typically obtained from 'computeRCN'.}

\item{g}{The number of clusters to create.}

\item{seed}{The random seed for reproducibility.}

\item{sort.by}{The cell type to sort clusters by (e.g., "CD8T").}

\item{sort.type}{The type of data to use for sorting: "freq" (relative frequencies) or "count" (counts).}

\item{sort.smpls}{The subset of samples to use for sorting: "all" (entire dataset) or "selected" (selected cells).}

\item{data.type}{The type of data to use for clustering: "ct_exp" (cell type and expression data) or "ct" (cell type data only).}

\item{extrapolate}{Whether to extrapolate clusters to the entire dataset (TRUE) or not (FALSE).}

\item{mc.cores}{The number of CPU cores to use for parallel processing.}
}
\value{
An updated 'frnn' object with clustering and sorting information.
}
\description{
This function clusters and sorts the Recurrent Cell Neighbors (RCN) for CyCIF or CyCIFStack objects.
It clusters cells based on their RCN profiles, sorts clusters based on the specified cell type,
and optionally extrapolates the clustering to the entire dataset.
}
\seealso{
\code{\link{computeRCN}}
}
