% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunUMAP.R
\name{RunUMAP}
\alias{RunUMAP}
\alias{RunUMAP,Cycif-method}
\alias{RunUMAP,CycifStack-method}
\title{Run visualization with dimensionality reduction (t-SNE and U-MAP) on CyCIF data.}
\usage{
RunUMAP(x, ...)

\S4method{RunUMAP}{Cycif}(
  x,
  norm_type = c("raw", "log_normalized", "logTh_normalized"),
  ld_name,
  ncells.per.smpl,
  used.abs,
  used.cts,
  ctype.full = FALSE,
  strict = TRUE,
  n_neighbors = 20,
  init.seed = 12345,
  ...
)

\S4method{RunUMAP}{CycifStack}(
  x,
  norm_type = c("raw", "log_normalized", "logTh_normalized"),
  ld_name,
  smpls,
  used.abs,
  used.cts,
  ctype.full = FALSE,
  strict = TRUE,
  ncells.per.smpl,
  n_neighbors = 20,
  init.seed = 12345,
  save.coords = FALSE,
  ...
)
}
\arguments{
\item{x}{A Cycif or CycifStack object.}

\item{...}{arguments passed to uwot::umap().}

\item{ncells.per.smpl}{A numeric scholar. The number of cells per sample to be set when a CycifStack is run.}

\item{used.abs}{A character vector containing a set of antibodies used for UMAP computation.}

\item{n_neighbors}{numeric. The number of neighbors, passed on to dimensionality reduction function.}

\item{init.seed}{initial seed to set for computing UMAP.}

\item{smpls}{Character vector containign a set of samples to be included inthe UMAP.}

\item{type}{character. It should be "raw" or "normalized", indicating which expression
values to use for the U-MAP.}

\item{n.cells}{numeric. The number of cells sampled from each CyCIF object.}
}
\description{
Run visualization with dimensionality reduction (t-SNE and U-MAP) on CyCIF data.
}
