% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis.R
\name{setDist}
\alias{setDist}
\alias{setDist,frNN-method}
\title{set distance to tumorBorder for frNN objects}
\usage{
setDist(x, ...)

\S4method{setDist}{frNN}(x, value)
}
\arguments{
\item{x}{A frNN object.}

\item{value}{A numeric vector specifying the distance to tumor border for each cell.}
}
\description{
set distance to tumorBorder for frNN objects
}
