% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{slidePlot}
\alias{slidePlot}
\alias{slidePlot,Cycif-method}
\title{Create a Slide Plot}
\usage{
slidePlot(x, ...)

\S4method{slidePlot}{Cycif}(
  x,
  pch = 20,
  cex = 2,
  bg.col,
  plot_type = c("dna", "exp", "cell_type", "custom"),
  custom_labs,
  ct_name = "default",
  strict = FALSE,
  ttl,
  leg.ttl,
  ab,
  uniq.cts,
  uniq.cols,
  draw.roi = FALSE,
  show.na = TRUE,
  na.col = "grey80",
  use_rois = TRUE,
  use.thres = TRUE,
  contour = FALSE,
  cont_nlevs = 3,
  ncells = 10000,
  trim_th = 0.01,
  legend = FALSE,
  legend.pos,
  mar = c(3, 3, 3, 3),
  roi.rec,
  ...
)
}
\arguments{
\item{x}{A Cycif object.}

\item{...}{Additional graphical parameters to customize the plot.}

\item{pch}{Plotting character for points (default is 20).}

\item{cex}{Plotting character size (default is 2).}

\item{bg.col}{Background color for points.}

\item{plot_type}{Type of slide plot to create: "dna" for DNA intensity, "exp" for protein expression,
"cell_type" for cell types, or "custom" for custom annotations.}

\item{custom_labs}{(For plot_type="custom") A factor vector of custom annotations to be used as labels.}

\item{ct_name}{The name of the cell type variable in the data.}

\item{strict}{(For plot_type="cell_type") If TRUE, cell type labels must exactly match those in the data.}

\item{ttl}{Title for the slide plot.}

\item{leg.ttl}{Legend title.}

\item{ab}{The name of the channel for plotting (e.g., "DNA1", "CD45", etc.).}

\item{uniq.cts}{(For plot_type="cell_type") A vector of unique cell type labels to be displayed.}

\item{uniq.cols}{(For plot_type="cell_type") A vector of unique colors corresponding to unique cell types.}

\item{draw.roi}{Should regions of interest (ROIs) be drawn on the slide plot? (default is FALSE)}

\item{show.na}{Should cells with missing values be displayed? (default is TRUE)}

\item{na.col}{Color for cells with missing values.}

\item{use_rois}{Should ROIs be used in the analysis? (default is TRUE)}

\item{use.thres}{Should thresholding be applied to the data? (default is TRUE)}

\item{contour}{Should contour lines be added to the plot? (default is FALSE)}

\item{cont_nlevs}{The number of contour levels (default is 3).}

\item{ncells}{Maximum number of cells to display (default is 1e4).}

\item{trim_th}{Trimming threshold for outlier removal (default is 1e-2).}

\item{legend}{Should a legend be displayed on the slide plot? (default is FALSE)}

\item{legend.pos}{Position of the legend on the slide plot (default is automatic).}

\item{mar}{Margins for the plot (default is c(3, 3, 3, 3)).}

\item{roi.rec}{A list specifying the region of interest (ROI) rectangle, including 'x' and 'y' ranges.}
}
\description{
This function generates a slide plot for visualizing different aspects of Cycif data, including DNA intensity,
protein expression, cell types, or custom annotations.
}
\details{
\itemize{
\item The \code{slidePlot} function generates a slide plot for visualizing Cycif data.
\item You can choose the type of slide plot using the \code{plot_type} parameter, which can be "dna" for DNA intensity,
"exp" for protein expression, "cell_type" for cell types, or "custom" for custom annotations.
\item The function allows you to display ROIs, customize color schemes, and control various plotting parameters.
}
}
