% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{used_cells}
\alias{used_cells}
\title{Get the used cells in a Cycif object}
\usage{
used_cells(x)
}
\arguments{
\item{x}{A Cycif object.}
}
\value{
A data frame containing the used cells in the Cycif object.
}
\description{
The \code{used_cells} function returns the used cells in a Cycif object.
}
\seealso{
\code{\link{nSamples}}: Get the number of samples in a CycifStack object.
\code{\link{names}}: Get the names of samples in a CycifStack object.
\code{\link{maxCycles}}: Get the maximum number of cycles in a CycifStack object.
\code{\link{within_rois}}: Check if samples in a CycifStack are within ROIs.
\code{\link{length}}: Equivalent to \code{nSamples} for CycifStack objects.
}
